% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect_sharepoint}
\alias{connect_sharepoint}
\title{Connect to SharePoint Site}
\usage{
connect_sharepoint(site_url, tenant = NULL)
}
\arguments{
\item{site_url}{Character string. The URL of the SharePoint site to connect to.}

\item{tenant}{Character string. Tenant name of the desired SharePoint site.}
}
\value{
A SharePoint site object that can be used with other SharePointR functions.
}
\description{
Establishes a connection to a Microsoft SharePoint site using the Microsoft365R package.
Authentication is handled automatically, using your Microsoft 365 credentials.
}
\details{
For the initial verification, you will be redirected to your default browser
to log in, or to request access from the sites administrator

If you encounter HTTP 401 (unauthorized) errors, you need to ensure you are able
to access the SharePoint site in your default browser.
}
\examples{
\dontrun{
# Connect to a SharePoint site
site <- connect_sharepoint("https://example.sharepoint.com/sites/mysite")

# Connect with explicit tenant
site <- connect_sharepoint("https://example.sharepoint.com/sites/mysite", "example")
}

}
