% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sharepoint_rmd.R
\name{convert_sharepoint_rmd}
\alias{convert_sharepoint_rmd}
\title{Convert standard markdown file to govspeak}
\usage{
convert_sharepoint_rmd(
  file_path,
  site,
  drive,
  remove_blocks = FALSE,
  sub_pattern = TRUE,
  page_break = "line"
)
}
\arguments{
\item{file_path}{Name and location of the *.md file you want to extract from sharepoint}

\item{site}{Name of sharepoint site you want to save a file to. Can be found in the URL of the Sharepoint location, e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail", the site name would be Rail}

\item{drive}{Name of the sharepoint drive you want to save the file to Can be found in the URL of the sharepoint location e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail/RailStats", the drive name is "RailStats". Can also be found using the list_sharepoint_drives() function in this package.}

\item{remove_blocks}{bool; decision to remove block elements from *.md file.
This includes code blocks and warnings}

\item{sub_pattern}{bool or vector; decision to increase hashed headers by
one level in *.md file.
  TRUE will substitute all, FALSE will substitute none, while a vector of
  the desired substitution levels allows individual headings to be modified
  as required.
  e.g. "#" will modify only first level headers.}

\item{page_break}{string; chooses what page breaks are converted to on
Whitehall.
If "line", page breaks are replaced with a horizontal rule. If "none"
they are replaced with a line break.
If "unchanged" they are not removed.}
}
\description{
Convert markdown file to Whitehall Publisher (GOV.UK) govspeak markdown
format
}
