% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{create_list_info}
\alias{create_list_info}
\title{Create listinfo object}
\usage{
create_list_info(
  template = c("documentLibrary", "genericList", "task", "survey", "announcements",
    "contacts"),
  content_types = NULL,
  hidden = NULL,
  call = caller_env()
)
}
\arguments{
\item{template}{Type of template to use in creating the list.}

\item{content_types}{Optional. Set \code{TRUE} for \code{contentTypesEnabled} to be enabled.}

\item{hidden}{Optional. Set \code{TRUE} for list to be hidden.}
}
\description{
Helper function to create a listInfo object for use internally by \code{\link[=create_sp_list]{create_sp_list()}}.
See: \url{https://learn.microsoft.com/en-us/graph/api/resources/listinfo?view=graph-rest-1.0}
}
\keyword{internal}
\keyword{lists}
