% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list_item.R
\name{create_sp_list_item}
\alias{create_sp_list_item}
\title{Alternate syntax for create list item}
\usage{
create_sp_list_item(
  ...,
  .sp_list = NULL,
  .fields = NULL,
  .keep_na = FALSE,
  .multi_fields = NULL
)
}
\arguments{
\item{...}{Ignored if .fields is supplied.}

\item{.sp_list}{A \code{ms_list} object.}

\item{.fields}{A named list or single row data frame.}

\item{.multi_fields}{Optional. Names of multi-value (Collection) fields, such
as multi-select choice columns, that should always be sent as an array with
an \code{"@odata.type"} annotation. If \code{NULL}, only fields with a length other
than 1 are treated as multi-value fields.}
}
\description{
\code{\link[=create_sp_list_item]{create_sp_list_item()}} is an alternative to the \code{create_item} method for
\code{Microsoft365R::ms_list} objects. This is used by \code{\link[=create_sp_list_items]{create_sp_list_items()}}
(instead of the \code{bulk_import} method) to better handle NA values that broken
for number columns.
}
\keyword{internal}
