% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyMoveDelete.R
\name{deleteSharepointItem}
\alias{deleteSharepointItem}
\title{Delete Sharepoint Item}
\usage{
deleteSharepointItem(team_name, folder_name, file_name)
}
\arguments{
\item{team_name}{name of team or drive}

\item{folder_name}{replace back slashes with forward slashes `/`. Feel free to keep spaces if present.}

\item{file_name}{include extension}
}
\value{

}
\description{
Delete Sharepoint Item
}
