% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_functions.R
\name{download_sharepoint_file}
\alias{download_sharepoint_file}
\title{Download a File from SharePoint}
\usage{
download_sharepoint_file(
  sharepoint_token,
  sharepoint_url,
  sharepoint_digest_value,
  sharepoint_path,
  sharepoint_file_name,
  out_path
)
}
\arguments{
\item{sharepoint_token}{A SharePoint token from get_sharepoint_token()}

\item{sharepoint_url}{A SharePoint url, e.g. kksky.sharepoint.com}

\item{sharepoint_digest_value}{A SharePoint digest value from get_sharepoint_digest_value()}

\item{sharepoint_path}{Path to the file, e.g. Shared Documents/test}

\item{sharepoint_file_name}{Name of the file to download, e.g. Dokument.docx}

\item{out_path}{Local path to write file to, e.g. C:/Dokuments}
}
\value{

}
\description{
Download a File from SharePoint
}
\examples{
#no example yet
}
