% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_user_info.R
\name{fmt_person_lookup_id_values}
\alias{fmt_person_lookup_id_values}
\title{Format a data frame with specified person columns to create or update
person/group columns in a SharePoint list}
\usage{
fmt_person_lookup_id_values(
  .data,
  cols,
  users_id = "EMail",
  site_users = NULL,
  sp_site = NULL
)
}
\description{
Format a data frame with specified person columns to create or update
person/group columns in a SharePoint list
}
\keyword{internal}
