% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Get IE proxy settings}
\usage{
get_config()
}
\value{
a httr::config request object. A list of settings.
}
\description{
This may alleviate some issues with work proxy settings. On the other hand
if you are not connected to the VPN and outside the network, then this
function will fail. The proxy script cannot be downloaded in these latter
cases (set \code{use_config = FALSE in function calls})
}
