% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cookies.R
\name{get_cookie}
\alias{get_cookie}
\title{Get cookies}
\usage{
get_cookie(base_site, token, get_config = F, config)
}
\arguments{
\item{base_site}{character string, the sharepoint base site address e.g. mysite.sharepoint.com '('https://' will be removed)}

\item{token}{the output from a successful \code{get_token()} call}

\item{get_config}{boolean, whether to use proxy script (from Internet Explorer settings).
Will fail if you are outside the network and not on the VPN}

\item{config}{httr::config request, a httr::config list}
}
\description{
Get cookies
}
