% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_functions.R
\name{get_sharepoint_digest_value}
\alias{get_sharepoint_digest_value}
\title{Generate a Digest Value}
\usage{
get_sharepoint_digest_value(sharepoint_token, sharepoint_url)
}
\arguments{
\item{sharepoint_token}{A SharePoint token from get_sharepoint_token()}

\item{sharepoint_url}{A SharePoint url, e.g. kksky.sharepoint.com}
}
\value{

}
\description{
Generate a Digest Value
}
\examples{
#no example yet
}
