% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_sharepoint_site}
\alias{get_sharepoint_site}
\title{Get the sharepoint site object}
\usage{
get_sharepoint_site(site)
}
\arguments{
\item{site}{Name of sharepoint site. Can be found in the URL of the Sharepoint location, e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail", the site name would be Rail}
}
\description{
Get the sharepoint site object
}
