% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_functions.R
\name{get_sharepoint_token}
\alias{get_sharepoint_token}
\title{Generate a SharePoint Token}
\usage{
get_sharepoint_token(
  client_id,
  client_secret,
  tenant_id,
  resource_id,
  site_domain
)
}
\arguments{
\item{client_id}{The ID of the client you created using _layouts/15/appregnew.aspx}

\item{client_secret}{The secret for the client you created using _layouts/15/appregnew.aspx}

\item{tenant_id}{Bearer realm="tenant ID" when calling curl -X GET -v -H "Authorization: Bearer" https://url_of_sharepoint_site/_vti_bin/client.svc/}

\item{resource_id}{client_id="resource ID" when calling curl -X GET -v -H "Authorization: Bearer" https://url_of_sharepoint_site/_vti_bin/client.svc/}

\item{site_domain}{URL for SharePoint domain, e.g. kksky.sharepoint.com}
}
\value{
A Token that can be used for calling the SharePoint API
}
\description{
Generate a SharePoint Token
}
