% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_sp_drive}
\alias{get_sp_drive}
\title{Get a Drive from a SharePoint Site}
\usage{
get_sp_drive(site, drive_name)
}
\arguments{
\item{site}{A SharePoint site object, as returned by \code{\link{connect_sharepoint}}.}

\item{drive_name}{Character string. The name of the document library/drive to access.}
}
\value{
A SharePoint drive object that can be used with read/write functions.
}
\description{
Retrieves a specific document library (drive) from a connected SharePoint site.
}
\examples{
\dontrun{
# Connect to a SharePoint site
site <- connect_sharepoint("https://example.sharepoint.com/sites/mysite")

# Get the "Documents" drive
drive <- get_sp_drive(site, "Documents")

# Get a different drive called "Shared Files"
drive <- get_sp_drive(site, "Shared Files")
}

}
