% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{get_sp_list}
\alias{get_sp_list}
\title{Get a SharePoint list}
\usage{
get_sp_list(
  list_name = NULL,
  list_id = NULL,
  ...,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  site = NULL,
  .default_drive_name = getOption("sharepointr.default_drive_name", "Documents"),
  items = TRUE,
  call = caller_env()
)
}
\arguments{
\item{list_name, list_id}{SharePoint List name or ID string.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_drive]{get_sp_drive}}
  \describe{
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{properties}}{If \code{TRUE}, return the drive properties instead of the
\code{ms_drive} object.}
  }}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL will be parsed and a site URL built using the tenant and site name if
found.}

\item{site}{A \code{ms_site} object. If site is supplied, \code{site_url} is ignored.}

\item{.default_drive_name}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}

\item{items}{If \code{TRUE} (default), use the \code{list_items} method to return a
data frame with all items in the specified list. If \code{FALSE}, return a
\code{ms_list} class object.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=get_sp_list]{get_sp_list()}} is a wrapper for the \code{get_list} and \code{list_items} methods.
This function is still under development and does not support the URL parsing
used by \code{\link[=get_sp_item]{get_sp_item()}}.
}
