% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get security token for logging in}
\usage{
get_token(base_site, username, password, get_config = F, config)
}
\arguments{
\item{base_site}{character string, the sharepoint base site address e.g. mysite.sharepoint.com '('https://' will be removed)}

\item{username}{character, your health email address}

\item{password}{character, your password}

\item{get_config}{boolean, whether to use proxy script (from Internet Explorer settings).
Will fail if you are outside the network and not on the VPN}

\item{config}{httr::config request, a httr::config list}
}
\description{
Get security token for logging in
}
