% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesAndDirectories.R
\name{getItemId}
\alias{getItemId}
\title{Title}
\usage{
getItemId(team_name, folder_path, filename)
}
\arguments{
\item{team_name}{name of team or drive}

\item{folder_path}{replace back slashes with forward slashes `/`. Feel free to keep spaces if present.}

\item{filename}{include extension}
}
\value{
item identification to be used for reading / writing / copying / moving / deleting
}
\description{
Title
}
