% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sp_url.R
\name{is_sp_url}
\alias{is_sp_url}
\alias{is_sp_site_url}
\alias{is_sp_doc_url}
\alias{is_sp_type_url}
\alias{is_sp_folder_url}
\title{Is x a SharePoint URL?}
\usage{
is_sp_url(x)

is_sp_site_url(x)

is_sp_doc_url(x)

is_sp_type_url(x, type = "w")

is_sp_folder_url(x)
}
\arguments{
\item{x}{Object to test if it is a URL.}

\item{type}{Type of URL to test against.}
}
\description{
Is x a SharePoint URL?
}
\keyword{internal}
