% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_drives.R
\name{list_sharepoint_objects}
\alias{list_sharepoint_objects}
\title{List the sharepoint objects available in the specified drive, useful for finding the full paths for objects}
\usage{
list_sharepoint_objects(site, drive, path = "")
}
\arguments{
\item{site}{Name of sharepoint site. Can be found in the URL of the Sharepoint location, e.g. for "departmentfortransportuk.sharepoint.com/sites/Rail", the site name would be Rail}

\item{drive}{Name of sharepoint drive you want to list objects associated with. This is case sensitive.}

\item{path}{By default, this function will list the objects at the top level of the drive. If you want to view lower-level items (e.g. inside a folder), you can specify the folder paths here. Default is blank}
}
\description{
List the sharepoint objects available in the specified drive, useful for finding the full paths for objects
}
