% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_drive.R
\name{list_sp_drives}
\alias{list_sp_drives}
\title{List drives for a SharePoint site}
\usage{
list_sp_drives(
  ...,
  site = NULL,
  filter = NULL,
  n = NULL,
  as_data_frame = TRUE,
  call = caller_env()
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link{get_sp_site}}
  \describe{
    \item{\code{site_url}}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}
    \item{\code{site_name,site_id}}{Site name or ID of the SharePoint site as an
alternative to the SharePoint site URL. Exactly one of  \code{site_url},
\code{site_name}, and \code{site_id} must be supplied.}
    \item{\code{refresh}}{If \code{TRUE}, get a new site even if the existing site is cached
as a local option. If \code{FALSE}, use the cached \code{ms_site} object.}
    \item{\code{cache}}{If \code{TRUE}, cache site to a file using \code{\link[=cache_sp_site]{cache_sp_site()}}.}
    \item{\code{cache_file}}{File name for cached drive or site. Default \code{NULL}.}
    \item{\code{overwrite}}{If \code{TRUE}, replace the existing cached object named by
\code{cache_file} with the new object. If \code{FALSE}, error if a cached file with
the same \code{cache_file} name already exists.}
    \item{\code{cache_dir}}{Cache directory. By default, uses an option
named "sharepointr.cache_dir". If "sharepointr.cache_dir" is not set, the
cache directory is set to \code{rappdirs::user_cache_dir("sharepointr")}.}
  }}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{filter}{Filter to apply to query}

\item{n}{Max number of drives to return}

\item{as_data_frame}{If \code{TRUE} (default), return list as a data frame.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A data frame with a list column of \code{ms_drive} objects if
\code{as_data_frame = TRUE} (default), or a list of \code{ms_drive} objects if
\code{FALSE}.
}
\description{
\code{\link[=list_sp_drives]{list_sp_drives()}} loads a SharePoint site uses the \code{list_drives} method to
returns a data frame with a list column or \code{ms_drive} objects or, if
\code{as_data_frame = FALSE}. This is helpful if a drive has been renamed and
can't easily be identified using a Drive URL alone.
}
\seealso{
\link[Microsoft365R:ms_site]{Microsoft365R::ms_site}
}
