% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesAndDirectories.R
\name{listFiles}
\alias{listFiles}
\title{List files in a folder}
\usage{
listFiles(team_name, folder_path)
}
\arguments{
\item{team_name}{name of team or drive}

\item{folder_path}{replace back slashes with forward slashes `/`}
}
\value{
`data.table` containing `File` (name of the file / subfolder) and `LastModified` date
}
\description{
List files in a folder
}
