% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyMoveDelete.R
\name{moveSharepointItem}
\alias{moveSharepointItem}
\title{Move Sharepoint Item}
\usage{
moveSharepointItem(
  read_team_name,
  read_folder,
  read_file,
  write_team_name,
  write_folder,
  write_file
)
}
\arguments{
\item{read_team_name}{acceptable values: `products`, `utils`}

\item{read_folder}{replace back slashes with forward slashes `/`. Feel free to keep spaces if present.}

\item{read_file}{include extension}

\item{write_team_name}{acceptable values: `products`, `utils`}

\item{write_folder}{replace back slashes with forward slashes `/`. Feel free to keep spaces if present.}

\item{write_file}{include extension}
}
\value{

}
\description{
Move Sharepoint Item
}
