% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ms_graph_obj_terms}
\alias{ms_graph_obj_terms}
\title{Shared general definitions for Microsoft Graph objects}
\arguments{
\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{plan}{A \code{ms_plan} object. If \code{plan} is supplied, \code{plan_title},
\code{plan_id}, and any additional parameters passed to \code{...} are ignored.}

\item{sp_list}{A \code{ms_list} object. If supplied, \code{list_name}, \code{list_id},
\code{site_url}, \code{site}, \code{drive_name}, \code{drive_id}, \code{drive}, and any additional
parameters passed to \code{...} are all ignored.}
}
\description{
Shared general definitions for Microsoft Graph objects
}
\keyword{internal}
