% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{ms_graph_terms}
\alias{ms_graph_terms}
\title{Shared general definitions for Microsoft Graph API parameters}
\arguments{
\item{filter}{A string with \href{https://learn.microsoft.com/en-us/graph/query-parameters?tabs=http#filter-parameter}{an OData expression}
apply as a filter to the results. Learn more in the \href{https://learn.microsoft.com/en-us/graph/filter-query-parameter}{Microsoft Graph API documentation}
on using filter query parameters.}

\item{n}{Maximum number of lists, plans, tasks, or other items to return.
Defaults to \code{NULL} which sets n to \code{Inf}.}
}
\description{
Shared general definitions for Microsoft Graph API parameters
}
\keyword{internal}
