% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_operations.R
\name{read_sp_csv}
\alias{read_sp_csv}
\title{Read in a CSV file from a SharePoint site}
\usage{
read_sp_csv(file_path, drive, ...)
}
\arguments{
\item{file_path}{Character string. The file path of the .csv file on the SharePoint site.)}

\item{drive}{A SharePoint drive object, as returned by \code{\link{get_sp_drive}}.}

\item{...}{Additional arguments passed to \code{\link[utils]{read.csv}}.}
}
\value{
A data frame object loaded into the users R environment.
}
\description{
Using an established connection to a SharePoint drive object, the function reads
a .csv data file into the users environment.
}
\details{
This function downloads the CSV file to a temporary location and then reads it into R
using the standard \code{\link[utils]{read.csv}} function. The temporary file is
automatically deleted after reading.

Ensure that your drive object is active and loaded in your environment.
}
\examples{
\dontrun{
# Connect to SharePoint and get drive
site <- connect_sharepoint("https://example.sharepoint.com/sites/mysite")
drive <- get_sp_drive(site, "Documents")

# Read CSV file
df <- read_sp_csv("folder_in_drive_object/data/data.csv", drive)

# With additional read.csv parameters
df2 <- read_sp_csv("data.csv", drive, sep = ";", header = FALSE)
}

}
