% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_sharepoint.R
\name{render_to_sharepoint}
\alias{render_to_sharepoint}
\title{Render an RMarkdown file and upload output to SharePoint}
\usage{
render_to_sharepoint(input_path, site, drive, dest_path, ...)
}
\arguments{
\item{input_path}{Path to the `.Rmd` file.}

\item{site}{SharePoint site name (e.g. "Rail").}

\item{drive}{SharePoint drive name (e.g. "Reports").}

\item{dest_path}{Path inside the SharePoint drive to upload to.}

\item{...}{Additional arguments passed to `rmarkdown::render()`.}
}
\description{
Render an RMarkdown file and upload output to SharePoint
}
