% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_get.R
\name{sharepoint_get}
\alias{sharepoint_get}
\title{Download file from SharePoint}
\usage{
sharepoint_get(path, token)
}
\arguments{
\item{path}{a character vector. Obtained by going online to SharePoint, right clicking on file, clicking 'details', clicking 'more details' in the resultant side pane, then clicking the 'copy' button right next to the "Path" header.}

\item{token}{a token from \code{sharepoint_token}}
}
\value{
a tempfile path pointing to the location of the file downloaded
}
\description{
Download file from SharePoint
}
