% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_put.R
\name{sharepoint_put}
\alias{sharepoint_put}
\title{Upload file to SharePoint}
\usage{
sharepoint_put(file, file_name = NULL, dest_path, token, overwrite = F)
}
\arguments{
\item{file}{path to file to be uploaded}

\item{file_name}{desired file name (with extension). Will be checked for uniqueness if overwrite = F}

\item{dest_path}{path to folder on SharePoint where file is to be uploaded}

\item{token}{AzureGraph token - see \code{sharepoint_token}}

\item{overwrite}{should the file overwrite that of one with a similar (see Details) name if it exists?}
}
\value{
an http response
}
\description{
Upload file to SharePoint
}
\details{
SharePoint API does not care about upper versus lower case.
As a matter of good practice, names should be distinct beyond just capitalization.
See the 'working-with-sharepoint-files.Rmd vignette for more details
}
