% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_token.R
\name{sharepoint_token}
\alias{sharepoint_token}
\title{Get SharePoint token}
\usage{
sharepoint_token()
}
\value{
a sharepoint token
}
\description{
A token is required whenever using the \code{sharepoint_get} or \code{sharepoint_put} function
}
\examples{
\dontrun{
token <- sharepoint_token()
}
}
