% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_get_pipeline.R
\name{sp_arrange}
\alias{sp_arrange}
\title{Arrange a SharePoint list}
\usage{
sp_arrange(table, ..., .arrange = NULL)
}
\arguments{
\item{table}{A SharePoint list connection
as returned by sp_list()}

\item{...}{Comma separated arrange commands}
}
\value{
Modfied SharePoint list connection
}
\description{
This method allows to arrange a SharePoint list
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_list = sp_list(sp_con, "yourList") \%>\% sp_arrange(Title, desc(column2))
}
