% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_driver.R
\name{sp_changeEscaping}
\alias{sp_changeEscaping}
\title{Change encoding of a SharePoint response}
\usage{
sp_changeEscaping(encoded)
}
\arguments{
\item{encoded}{An encoded SharePoint response}
}
\value{
SharePoint response with changed encoding
}
\description{
Change encoding of a SharePoint response
}
