% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_get_pipeline.R
\name{sp_collect}
\alias{sp_collect}
\title{Collect a SharePoint list}
\usage{
sp_collect(table, n = Inf, skip = NULL, expand = T, verbose = T)
}
\arguments{
\item{table}{A SharePoint list connection
as returned by sp_list()}

\item{n}{Number of rows to collect}

\item{skip}{Number of rows to skip}

\item{expand}{Follow deferred links}

\item{verbose}{Print status information}
}
\value{
Request result as data frame
}
\description{
This method allows to collect the results
of a SharePoint list request
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_list = sp_list(sp_con, "yourList") \%>\% sp_select(Title, `Column with whitespaces`) \%>\% sp_collect()
}
