% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_driver.R
\name{sp_connection}
\alias{sp_connection}
\title{Establish a connection to SharePoint}
\usage{
sp_connection(Address, Username = NULL, Password = NULL,
  credentialFile = NULL, Office365 = T)
}
\arguments{
\item{Address}{Address of the SharePoint instance to be contacted.
For example https://example.sharepoint.com}

\item{Username}{The username to use for the authentication.
If not given, a credential file must be provided.
For example: john.doe@example.com}

\item{Password}{The password to use for the authentication.
If not given, a credential file must be provided.}

\item{credentialFile}{If no username or password is provided,
this file will be used for reading the credentials from.
The file must be YAML formatted, containing the keys Username
and Password.}

\item{Office365}{True (default) if a connection to SharePoint online
shall be established, False if a connection to SharePoint server
shall be established.}
}
\description{
This method allows to establish and save a connection to
SharePoint online or a SharePoint server. The connection
properties are therefore saved in a connection object
to be later used for the request
}
\examples{
sp_connection("https://your.sharepoint.server/subpage1/", "YourUsername", "YourPassword", Office365 = F)
sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
}
