% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_create_folder.R
\name{sp_create_folder}
\alias{sp_create_folder}
\title{Create a folder on SharePoint}
\usage{
sp_create_folder(connection, rurl)
}
\arguments{
\item{connection}{a SharePoint connection}

\item{rurl}{character, sub-directory of the SharePoint site}
}
\description{
Create a folder on SharePoint
}
