% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_drive.R
\name{sp_drive}
\alias{sp_drive}
\alias{get_sp_drive}
\alias{cache_sp_drive}
\title{Get a SharePoint drive or set a default SharePoint drive}
\usage{
get_sp_drive(
  drive_name = NULL,
  drive_id = NULL,
  drive_url = NULL,
  properties = FALSE,
  ...,
  site_url = NULL,
  site = NULL,
  default_drive_name = getOption("sharepointr.default_drive_name", "Documents"),
  cache = getOption("sharepointr.cache", FALSE),
  refresh = getOption("sharepointr.refresh", TRUE),
  overwrite = FALSE,
  cache_file = NULL,
  call = caller_env()
)

cache_sp_drive(
  ...,
  drive = NULL,
  cache_file = getOption("sharepointr.cache_file_drive", "sp_drive.rds"),
  overwrite = FALSE,
  call = caller_env()
)
}
\arguments{
\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive_url}{A SharePoint Drive URL to parse for a Drive name and other
information. If \code{drive_name} is a URL, it is used as \code{drive_url}.}

\item{properties}{If \code{TRUE}, return the drive properties instead of the
\code{ms_drive} object. Defaults to \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link{get_sp_site}}
  \describe{
    \item{\code{site_name,site_id}}{Site name or ID of the SharePoint site as an
alternative to the SharePoint site URL. Exactly one of  \code{site_url},
\code{site_name}, and \code{site_id} must be supplied.}
    \item{\code{cache_dir}}{Cache directory. By default, uses an option
named "sharepointr.cache_dir". If "sharepointr.cache_dir" is not set, the
cache directory is set to \code{rappdirs::user_cache_dir("sharepointr")}.}
  }}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{default_drive_name}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}

\item{cache}{If \code{TRUE}, cache drive to a file using \code{\link[=cache_sp_drive]{cache_sp_drive()}}.}

\item{refresh}{If \code{TRUE}, get a new drive even if the existing drive is
cached as a local option. If \code{FALSE}, use the cached \code{ms_drive} object if
it exists.}

\item{overwrite}{If \code{TRUE}, replace the existing cached object named by
\code{cache_file} with the new object. If \code{FALSE}, error if a cached file with
the same \code{cache_file} name already exists.}

\item{cache_file}{File name for cached drive or site. Default \code{NULL}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{drive}{A \code{ms_drive} object. If \code{drive} is supplied, \code{drive_name} and
\code{drive_id} are ignored.}
}
\value{
For \code{\link[=get_sp_drive]{get_sp_drive()}}, a \code{ms_drive} object, or the drive
properties as a named list if \code{properties = TRUE}. For
\code{\link[=cache_sp_drive]{cache_sp_drive()}}, invisibly returns \code{NULL}.
}
\description{
\code{\link[=get_sp_drive]{get_sp_drive()}} wraps the \code{get_drive} method returns a \code{ms_drive} object.

\code{\link[=cache_sp_drive]{cache_sp_drive()}} allows you to cache a default SharePoint drive for use by
other functions. Additional parameters in \code{...} are passed to
\code{\link[=get_sp_drive]{get_sp_drive()}}.
}
\seealso{
\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive}
}
\keyword{drives}
