% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_get_pipeline.R
\name{sp_explain}
\alias{sp_explain}
\title{Explain a SharePoint list request}
\usage{
sp_explain(table)
}
\arguments{
\item{table}{A SharePoint list connection}
}
\value{
A SharePoint list connection
}
\description{
This method allows to explain a SharePoint
list request
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_list = sp_list(sp_con, "yourList") \%>\% sp_select(Title, `Column with whitespaces`) \%>\% sp_explain
}
