% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_get_pipeline.R
\name{sp_filter}
\alias{sp_filter}
\title{Apply filter operations on a SharePoint list}
\usage{
sp_filter(table, ..., .filter = NULL)
}
\arguments{
\item{table}{A SharePoint list connection
as returned by sp_list()}

\item{...}{Comma separated filter commands
white spaces in filters can be escaped using ``.
For quoting, !! is used.}
}
\value{
Modfied SharePoint list connection
}
\description{
This method allows to apply filter operations
on a SharePoint list connection. The syntax is
inspired by dplyr
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_list = sp_list(sp_con, "yourList") \%>\% sp_filter(Title == "yourTitle")
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_list = sp_list(sp_con, "yourList") \%>\% sp_filter(`Your column` == "yourTitle")
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_list = sp_list(sp_con, "yourList") \%>\% sp_filter(Title == !!yourVariable)
}
