% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_get_file.R
\name{sp_get_file}
\alias{sp_get_file}
\title{Download a file from SharePoint using a HTTP GET request}
\usage{
sp_get_file(connection, rurl, destfile)
}
\arguments{
\item{connection}{a sp_connection object, that from a successful \code{sp_connection()} call}

\item{rurl}{a character string, the relative url (rurl) of the file you want to download}

\item{destfile}{a character string, the file name (and possibly file path) you want to save to. If missing
the file name from the connection will be used and file saved to the working directory.}
}
\value{
A message on whether the download was successful or not.
}
\description{
Download a file from SharePoint using a HTTP GET request
}
