% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_get.R
\name{sp_getListMetadata}
\alias{sp_getListMetadata}
\title{Get metadata of a SharePoint list}
\usage{
sp_getListMetadata(con, listName = NULL, listID = NULL, raw = F)
}
\arguments{
\item{con}{A SharePoint connection}

\item{listName}{Name of the SharePoint list}

\item{raw}{Return response raw formatted. Default is
FALSE}
}
\value{
SharePoint response as list (raw = TRUE) or
the decoded response content as list (raw = FALSE).
}
\description{
This method lists all available metadata for a
SharePoint list.
}
