% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_get.R
\name{sp_getLists}
\alias{sp_getLists}
\title{List available SharePoint lists}
\usage{
sp_getLists(con, raw = F)
}
\arguments{
\item{con}{A SharePoint connection returned
by sp_connection()}

\item{raw}{Return response raw formatted. Default is
FALSE}
}
\value{
SharePoint response as list (raw = TRUE) or
the available lists as character vector (raw = FALSE).
}
\description{
This method lists all available lists for a
SharePoint connection.
}
