% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_driver.R
\name{sp_getRequestDigest}
\alias{sp_getRequestDigest}
\title{Get a SharePoint request digest}
\usage{
sp_getRequestDigest(con)
}
\arguments{
\item{con}{A SharePoint connection returned
by sp_connection()}
}
\value{
Request digest as string
}
\description{
Request and receive a SharePoint request
digest for further editing
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
digest = sp_getRequestDigest(sp_con)
}
