% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list_contents.R
\name{sp_list_contents}
\alias{sp_list_contents}
\title{List the contents of a SharePoint directory}
\usage{
sp_list_contents(connection, rurl, recursive = F)
}
\arguments{
\item{connection}{a sp_connection object, that from a successful \code{sp_connection()} call}

\item{rurl}{a character string, the relative url (rurl) of the folder you want to list contents}

\item{recursive}{boolean, recursively search the directories. Not implemented yet.}

\item{use_config}{boolean, whether to use proxy script (from Internet Explorer settings).
Will fail if you are outside the network and not on the VPN}
}
\value{
A data.frame with type, name and relative url information.
}
\description{
List the contents of a SharePoint directory
}
