% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list_files.R
\name{sp_list_files}
\alias{sp_list_files}
\title{Get full file information from an SharePoint site}
\usage{
sp_list_files(connection, rurl)
}
\arguments{
\item{connection}{a SharePoint connection}

\item{rurl}{character, sub-directory of the SharePoint site}
}
\value{
a data.frame of the file information from the sub-directory
}
\description{
Get full file information from an SharePoint site
}
