% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list_folders.R
\name{sp_list_folders}
\alias{sp_list_folders}
\title{Get full folder information from an SharePoint site}
\usage{
sp_list_folders(connection, sub_dir)
}
\arguments{
\item{connection}{a SharePoint connection}

\item{sub_dir}{character, sub-directory of the SharePoint site}
}
\value{
a data.frame of the folder information from the sub-directory
}
\description{
Get full folder information from an SharePoint site
}
