% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_operations.R
\name{sp_load}
\alias{sp_load}
\title{Load R Data Files from SharePoint}
\usage{
sp_load(file_path, drive)
}
\arguments{
\item{file_path}{Character string. The path to the R data file on the SharePoint drive.}

\item{drive}{A SharePoint drive object, as returned by \code{\link{get_sp_drive}}.}
}
\value{
The R objects in the data file are loaded into the global environment.
}
\description{
Loads an R data file (.RData or .rda) from SharePoint into the global environment.
}
\details{
This function downloads the R data file to a temporary location and then loads it into
the global environment using the standard \code{\link[base]{load}} function.
The temporary file is automatically deleted after loading.

The function supports both .RData and .rda file extensions, which are standard
formats for saving R objects.
}
\examples{
\dontrun{
# Connect to SharePoint and get drive
site <- connect_sharepoint("https://example.sharepoint.com/sites/mysite")
drive <- get_sp_drive(site, "Documents")

# Load R data file
sp_load("models/regression_model.RData", drive)

# After running this, the objects saved in regression_model.RData
# will be available in your global environment
}

}
