% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_plan.R
\name{sp_plan}
\alias{sp_plan}
\alias{get_sp_plan}
\alias{list_sp_plans}
\title{Get a SharePoint plan or list SharePoint plans}
\usage{
get_sp_plan(
  plan_title = NULL,
  plan_id = NULL,
  ...,
  site = NULL,
  site_url = NULL,
  as_data_frame = FALSE,
  call = caller_env()
)

list_sp_plans(
  ...,
  filter = NULL,
  n = NULL,
  site = NULL,
  as_data_frame = TRUE,
  call = caller_env()
)
}
\arguments{
\item{plan_title, plan_id}{Planner title or ID. Exactly one of the two
arguments must be supplied.}

\item{...}{Additional arguments passed to \code{\link[=get_sp_group]{get_sp_group()}}.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{as_data_frame}{If \code{TRUE} (default for \code{\link[=list_sp_plans]{list_sp_plans()}}), return a data
frame with the plan title, id, creation date/time, and owner ID with a list
column named "ms_plan" containing \code{ms_plan} objects. If \code{FALSE} (default
\code{\link[=get_sp_plan]{get_sp_plan()}}), return a \code{ms_plan} object or list of \code{ms_plan} objects.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{filter}{A string with \href{https://learn.microsoft.com/en-us/graph/query-parameters?tabs=http#filter-parameter}{an OData expression}
apply as a filter to the results. Learn more in the \href{https://learn.microsoft.com/en-us/graph/filter-query-parameter}{Microsoft Graph API documentation}
on using filter query parameters.}

\item{n}{Maximum number of lists, plans, tasks, or other items to return.
Defaults to \code{NULL} which sets n to \code{Inf}.}
}
\value{
For \code{\link[=get_sp_plan]{get_sp_plan()}}, a \code{ms_plan} class object or a 1 row data frame
with a "ms_plan" column.

For \code{\link[=list_sp_plans]{list_sp_plans()}}, A list of \code{ms_plan} class objects or a data
frame with a list column named "ms_plan".
}
\description{
\code{\link[=get_sp_plan]{get_sp_plan()}} returns a single \code{ms_plan} object using the \code{get_plan}
method. \code{\link[=list_sp_plans]{list_sp_plans()}} returns a data frame with plan properties or a list
of \code{ms_plan} objects.
}
\seealso{
\link[Microsoft365R:ms_plan]{Microsoft365R::ms_plan}
}
\keyword{plans}
