% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_post_file.R
\name{sp_post_file}
\alias{sp_post_file}
\title{Upload a file to SharePoint using a HTTP POST request}
\usage{
sp_post_file(connection, rurl, file)
}
\arguments{
\item{connection}{a sp_connection object, that from a successful \code{sp_connection()} call}

\item{rurl}{a character string, the relative url (rurl) of the folder you want to upload to}

\item{file}{a character string, the file you want to upload from your system
Will fail if you are outside the network and not on the VPN}
}
\value{
A message on whether the upload was successful or not.
}
\description{
Upload a file to SharePoint using a HTTP POST request
}
