% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_qget.R
\name{sp_qget}
\alias{sp_qget}
\title{Wrapper to download a sharepoint file}
\usage{
sp_qget(url, destfile, username, password, get_config = F, config)
}
\arguments{
\item{url}{character, full URL of sharepoint file wanted}

\item{destfile}{character, the file name to save the file to}

\item{username}{character, your health email address. Using a keyring will help hide your password in workflows.}

\item{password}{character, your password}

\item{get_config}{boolean, whether to use proxy script (from Internet Explorer settings).
Will fail if you are outside the network and not on the VPN}

\item{config}{httr::config request, a httr::config list}
}
\description{
Wrapper to download a sharepoint file
}
\examples{

\dontrun{

url = "https://n...h.sharepoint.com/sites/group_i_am_part_of/folder_1/file.xlsx"

get_sharepoint_file(url = url,
                    destfile = 'file.xlsx',
                    username = 'me@this.place.com',
                    password = keyring::key_get('mail'))

}
}
