% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{sp_read_rds}
\alias{sp_read_rds}
\title{Load RDS Files from SharePoint}
\usage{
sp_read_rds(file_path, drive)
}
\arguments{
\item{file_path}{Character string. The path to the R data file on the SharePoint drive.}

\item{drive}{A SharePoint drive object, as returned by \code{\link{get_sp_drive}}.}
}
\value{
The R objects in the data file are loaded into the global environment.
}
\description{
Loads an RDS data file (.rds) from SharePoint into the global environment.
}
\details{
This function downloads the RDS file to a temporary location and then loads it into
the global environment using the standard \code{\link[base]{load}} function.
The temporary file is automatically deleted after loading.
}
\examples{
\dontrun{
# Connect to SharePoint and get drive
site <- connect_sharepoint("https://example.sharepoint.com/sites/mysite")
drive <- get_sp_drive(site, "Documents")

# Load R data file
sp_read_rds("models/regression_model.rds", drive)

# After running this, the objects saved in regression_model.RData
# will be available in your global environment
}

}
