% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_driver.R
\name{sp_request}
\alias{sp_request}
\title{Send a request to the SharePoint API}
\usage{
sp_request(con, request, verb = "GET", json = T, body = NULL)
}
\arguments{
\item{con}{A SharePoint connection returned
by sp_connection()}

\item{request}{A request URL. Can be either a full URL
or the last part of the request URL which will be
concatenated to the API-URL of the SharePoint connection}

\item{verb}{Method to use for the request. Either GET
(default) or POST.}

\item{json}{retreive data using JSON formatting}

\item{body}{Body of the request (only for POST).}
}
\value{
The response as list
}
\description{
Send a request via GET or POST to the
SharePoint RESTful API
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
lists = sp_request(sp_con, "lists", verb = "GET", json = T, body = NULL)
}
