% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_operations.R
\name{sp_save}
\alias{sp_save}
\title{Save R Objects to an R Data File on SharePoint}
\usage{
sp_save(objects, file_path, drive, overwrite = FALSE)
}
\arguments{
\item{objects}{Character vector. Names of objects in the global environment to save.}

\item{file_path}{Character string. The path where the file should be written on the SharePoint drive.}

\item{drive}{A SharePoint drive object, as returned by \code{\link{get_sp_drive}}.}

\item{overwrite}{Logical. If TRUE, overwrites the file if it already exists. Default is FALSE.}
}
\value{
Returns console message indicating status of export.
}
\description{
Saves specified R objects to an .RData file and uploads it to SharePoint.
}
\details{
This function saves specified R objects to a temporary .RData file and then uploads it to
SharePoint. The temporary file is automatically deleted after uploading.

Folders in the file_path must already exist. The function will not create new folders.
If you want to save to a nested folder structure, ensure all folders exist
on the SharePoint drive first.

Only objects that exist in the global environment can be saved. If any object
in the list doesn't exist, an error will be returned.
}
\examples{
\dontrun{
# Connect to SharePoint and get drive
site <- connect_sharepoint("https://example.sharepoint.com/sites/mysite")
drive <- get_sp_drive(site, "Documents")

# Create some objects to save
model <- lm(mpg ~ wt + hp, data = mtcars)
summary_stats <- summary(model)

# Save the objects to SharePoint
sp_save(c("model", "summary_stats"), "models/regression_results.RData", drive)

# Overwrite an existing file
sp_save(c("model", "summary_stats"), "models/regression_results.RData", drive, overwrite = TRUE)
}

}
