% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_url_parse.R
\name{sp_url_parse}
\alias{sp_url_parse}
\alias{sp_url_parse_hostname}
\alias{sp_url_parse_path}
\alias{sp_url_parse_query}
\title{Parse a SharePoint URL using \code{httr2::url_parse}}
\usage{
sp_url_parse(url, call = caller_env())

sp_url_parse_hostname(hostname, tenant = "[a-zA-Z0-9.-]+", scheme = "https")

sp_url_parse_path(
  path,
  url_type = "w|x|p|o|b|t|i|v|f|u|li",
  permissions = "r|s|t|u|g",
  drive_name_prefix = "Shared ",
  default_drive_name = "Documents"
)

sp_url_parse_query(query)
}
\description{
\code{\link[=sp_url_parse]{sp_url_parse()}} parses a URL into a named list of parts.

\code{\link[=sp_url_parse_hostname]{sp_url_parse_hostname()}} parses the hostname into a tenant and base_url.

\code{\link[=sp_url_parse_path]{sp_url_parse_path()}} parses the path into a URL type, permissions, drive
name, file path, and file name. \code{path} must already be decoded (as returned
by \code{\link[httr2:url_parse]{httr2::url_parse()}}), e.g. \code{"/:f:/r/sites/site-name/Shared Documents/Folder"}.

\code{\link[=sp_url_parse_query]{sp_url_parse_query()}} parses the item ID from a query.
}
\details{
Types of SharePoint URLs

SharePoint site URL:

https://[tenant].sharepoint.com/sites/[site name]

SharePoint document editor URL:

https://[tenant].sharepoint.com/:[url
type]:/[permissions]/sites/[site
name]/_layouts/15/Doc.aspx?sourcedoc=\{[item id]\}&file=[file
name]&[additional query parameters]

SharePoint List URL:

https://[tenant].sharepoint.com/sites/[site name]/Lists/[list
name]/AllItems.aspx?env=WebViewList

https://[tenant].sharepoint.com/sites/[site name]/Lists/[list
name]

https://[tenant].sharepoint.com/:l:/r/sites/[site name]/Lists/[list
name]

SharePoint folder URL:

https://[tenant].sharepoint.com/:[url type]:/[permissions]/sites/[site
name]/[drive name (with possible "Shared" prefix)]/[folder path]/[folder
name]/?[additional query parameters]

SharePoint Planner URL:

https://tasks.office.com/[tenant].onmicrosoft.com/en-US/Home/Planner/#/
plantaskboard?groupId=[Group ID]&planId=[Plan ID] (former)

https://planner.cloud.microsoft/webui/v1/plan/[Plan ID]?tid=[tenant]
(current)

SharePoint Document Library (or Drive) URL:

https://[tenant].sharepoint.com/sites/[tenant]/[drive name (with
possible "Shared" prefix)]/Forms/AllItems.aspx
}
\keyword{internal}
