% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_users.R
\name{sp_users}
\alias{sp_users}
\title{Retrieve SharePoint users}
\usage{
sp_users(con, ..., .filter = NULL)
}
\arguments{
\item{con}{A SharePoint connection returned by
sp_connection()}

\item{...}{Filters}

\item{filter}{A Character string containing
valid SharePoint filters (for example:
LoginName eq Username1)}
}
\value{
User information as a dataframe
}
\description{
Retrieve all SharePoint users of the given site
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
user = sp_users(sp_con, LoginName = "Username1")
}
