% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_post.R
\name{sp_writeListData}
\alias{sp_writeListData}
\title{Write data to a SharePoint list}
\usage{
sp_writeListData(con, listName = NULL, listID = NULL, data)
}
\arguments{
\item{con}{A SharePoint connection returned
by sp_connection()}

\item{listName}{Name of the SharePoint list
to write to}

\item{data}{List or dataframe of the data to
write into the SharePoint list}
}
\description{
This method allows to write data to a SharePoint
list. Therefore, type checks and lookups will be
performed (see details).
}
\details{
In order to upload the data correctly,
the following data types are casted automatically
\enumerate{
  \item Text columns
  \item Numeric columns
  \item User names
  \item DateTime columns
  \item Lookup columns linking to another table
}
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_writeListData(sp_con, "yourList", list(column1 = "content1", column2 = 2, column3 = 1.23, user = "Username"))
}
