% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_functions.R
\name{upload_file_to_sharepoint}
\alias{upload_file_to_sharepoint}
\title{Upload a File to SharePoint}
\usage{
upload_file_to_sharepoint(
  sharepoint_token,
  sharepoint_url,
  sharepoint_digest_value,
  sharepoint_path,
  sharepoint_file_name,
  file_path
)
}
\arguments{
\item{sharepoint_token}{A SharePoint token from get_sharepoint_token()}

\item{sharepoint_url}{A SharePoint url, e.g. kksky.sharepoint.com}

\item{sharepoint_digest_value}{A SharePoint digest value from get_sharepoint_digest_value()}

\item{sharepoint_path}{Path to the file, e.g. Shared Documents/test}

\item{sharepoint_file_name}{Name of the file in SharePoint, e.g. Dokument.docx}

\item{file_path}{Path to the file you want to upload, e.g. C:/Dokument.docx}
}
\value{

}
\description{
Upload a File to SharePoint
}
