% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sharepoint.R
\name{upload_sp_item}
\alias{upload_sp_item}
\alias{upload_sp_items}
\title{Upload a file or folder to a SharePoint Drive}
\usage{
upload_sp_item(
  file = NULL,
  dest = NULL,
  ...,
  src = NULL,
  overwrite = FALSE,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  blocksize = 327680000,
  recursive = FALSE,
  parallel = FALSE,
  call = caller_env()
)

upload_sp_items(file = NULL, dest, ..., src = NULL, call = caller_env())
}
\arguments{
\item{file}{Path for file or directory to upload. Optional if \code{src} is
supplied.}

\item{dest}{Destination on SharePoint for file to upload. Must be a
\code{ms_drive_item} for a folder, a SharePoint folder URL, or a drive path. If a
\code{ms_drive_item} is provided and \code{drive_name} is \code{NULL}, the item URL is
used to determine the drive. If \code{NULL}, the top level of the provided drive
is used as the destination. If file is a directory, \code{dest} must be an
\emph{existing} directory.}

\item{...}{Additional parameters passed to \code{\link[=get_sp_site]{get_sp_site()}} or
\code{\link[Microsoft365R:get_sharepoint_site]{Microsoft365R::get_sharepoint_site()}}.}

\item{src}{Data source path passed to \code{upload_folder} or \code{upload_file}
method. Defaults to \code{NULL} and set to use file value by default.}

\item{overwrite}{If \code{FALSE} (default), error if an item with the name
specified in \code{file} or \code{src} already exists at the specified destination.
If \code{TRUE}, overwrite any existing items with the same name. The latter is
the default for the \code{upload_file} method.}

\item{drive_name, drive_id}{SharePoint drive name or ID.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{blocksize, recursive, parallel}{Additional parameters passed to
\code{upload_folder} or \code{upload_file} method for \code{ms_drive} objects.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Invisibly returns the destination path (\code{dest}) for the uploaded
file or directory.
}
\description{
\code{\link[=upload_sp_item]{upload_sp_item()}} wraps the \code{upload_folder} and \code{upload_file} method for
\code{ms_drive} objects.
}
