% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{write_excel_to_sharepoint}
\alias{write_excel_to_sharepoint}
\title{Write excel file to a Sharepoint location using a template}
\usage{
write_excel_to_sharepoint(
  read_team_name,
  read_folder,
  read_file,
  write_team_name,
  write_folder,
  write_file,
  dta,
  preserve_character = FALSE
)
}
\arguments{
\item{read_team_name}{name of team or drive}

\item{read_folder}{replace back slashes with forward slashes `/`. Feel free to keep spaces if present.}

\item{read_file}{include extension}

\item{write_team_name}{name of team or drive}

\item{write_folder}{replace back slashes with forward slashes `/`. Feel free to keep spaces if present.}

\item{write_file}{include extension}

\item{dta}{`data.frame` or `data.table`}

\item{preserve_character}{if false, this will recognize `currency`, `percentage`, `character` class.
These classes will be converted to the respective excel format thusly. The rest will be `General`
If true, everything is `General`.}
}
\value{

}
\description{
Write excel file to a Sharepoint location using a template
}
