% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sharepoint.R
\name{write_sharepoint}
\alias{write_sharepoint}
\title{Write an object to file and upload file to SharePoint}
\usage{
write_sharepoint(
  x,
  file,
  dest,
  ...,
  .f = NULL,
  new_path = tempdir(),
  overwrite = FALSE,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  site_name = NULL,
  site_id = NULL,
  site = NULL,
  blocksize = 327680000,
  call = caller_env()
)
}
\arguments{
\item{x}{Object to write to file and upload to SharePoint. \code{sf}, \code{rdocx},
\code{rpptx}, \code{rxlsx}, and \code{data.frame} objects are saved with the functions
noted in the description. All other object types are saved as \code{rds}
outputs.}

\item{file}{File to write to. Passed to \code{file} parameter for
\code{\link[readr:write_csv]{readr::write_csv()}} or \code{\link[readr:write_rds]{readr::write_rds()}}, \code{dsn} for \code{sf::write_sf()},
or \code{target} for \code{print()} (when working with \code{{officer}} class objects).}

\item{dest}{Destination on SharePoint for file to upload. Must be a
\code{ms_drive_item} for a folder, a SharePoint folder URL, or a drive path. If
a \code{ms_drive_item} is provided and \code{drive_name} is NULL, the item URL is used
to determine the drive. If \code{NULL}, the top level of the provided drive is
used as the destination. If file is a directory, \code{dest} must be an \emph{existing}
directory.}

\item{...}{Additional parameters passed to write function.}

\item{.f}{Optional function to write the input data to disk before uploading
file to SharePoint.}

\item{new_path}{Path to write file to. Defaults to \code{\link[=tempdir]{tempdir()}}}

\item{overwrite}{If \code{FALSE} (default), error if an item with the name
specified in \code{file} or \code{src} already exists at the specified destination.
If \code{TRUE}, overwrite any existing items with the same name. The latter is
the default for the \code{upload_file} method.}

\item{drive_name, drive_id}{SharePoint drive name or ID.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site_name, site_id}{Site name or ID of the SharePoint site as an
alternative to the SharePoint site URL. Exactly one of  \code{site_url},
\code{site_name}, and \code{site_id} must be supplied.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{blocksize}{Additional parameter passed to \code{upload_folder} or
\code{upload_file} method for \code{ms_drive} objects.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Invisibly returns the input object x.
}
\description{
\code{\link[=write_sharepoint]{write_sharepoint()}} uses \code{\link[=upload_sp_item]{upload_sp_item()}} to upload an object created
using:
\itemize{
\item \code{\link[sf:write_sf]{sf::write_sf()}} if x is an \code{sf} object
\item \code{\link[readr:write_csv]{readr::write_csv()}} if x is a \code{data.frame} and file does not include a xlsx extension
\item \code{\link[openxlsx2:write_xlsx]{openxlsx2::write_xlsx()}} if x is a \code{data.frame} file does include a xlsx extension
\item The \code{\link[=print]{print()}} method from \code{{officer}} if x is a \code{rdocx}, \code{rpptx}, or \code{rxlsx} object
\item \code{\link[readr:write_rds]{readr::write_rds()}} if x is any other class
}
}
