% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{writeColumn}
\alias{writeColumn}
\title{Write a single column into exisiting excel file in Sharepoint}
\usage{
writeColumn(
  drive_id,
  write_file_id,
  sheetid,
  session_id,
  dta,
  column_id,
  column_name,
  preserve_class
)
}
\arguments{
\item{preserve_class}{}
}
\value{

}
\description{
Write a single column into exisiting excel file in Sharepoint
}
