% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureR_config.R
\name{AzureR_config}
\alias{AzureR_config}
\alias{AzureR_config_ls}
\alias{AzureR_config_delete}
\title{List or delete files from the AzureR configuration directory}
\usage{
AzureR_config_ls(path = NULL, glob = "*.json")

AzureR_config_delete(path = NULL, filename = "graph_logins.json")
}
\arguments{
\item{path}{Path to configuration directory for AzureR package where the JSON
file for graph_logins is stored. If \code{NULL}, path is set with the base
folder for \code{rappdirs::user_cache_dir("AzureR")} (for Windows) or
\code{rappdirs::user_config_dir("AzureR")} (for non-Windows).}

\item{glob}{A wildcard aka globbing pattern (e.g. \verb{*.csv}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{filename}{Filename to delete from configuration directory. Defaults to
"graph_logins.json". Set to \code{NULL} if path contains a file name.}
}
\description{
\code{\link[=AzureR_config_ls]{AzureR_config_ls()}} uses \code{\link[rappdirs:user_config_dir]{rappdirs::user_config_dir()}} and \code{\link[fs:dir_ls]{fs::dir_ls()}} to
list files from the AzureR configuration directory. \code{\link[=AzureR_config_delete]{AzureR_config_delete()}}
uses \code{\link[fs:file_delete]{fs::file_delete()}} to remove the "graph_logins.json" configuration file
if needed. Use this function with caution but it may be an option to address
a "Unable to refresh token" error.
}
