% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_cache.R
\name{cache_ms_obj}
\alias{cache_ms_obj}
\title{Cache a \code{ms_object} class object to a file}
\usage{
cache_ms_obj(
  x,
  cache_file,
  cache_dir = NULL,
  what = "ms_site",
  overwrite = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{cache_file}{File name for cached file. Required.}

\item{cache_dir}{Cache directory. By default, uses an environmental variable
named "sharepointr.cache_dir". If "sharepointr.cache_dir" is not set, the
cache directory is set to \code{rappdirs::user_cache_dir("sharepointr")}.}

\item{overwrite}{If \code{TRUE}, replace the existing cached object named by
\code{cache_file} with the new object. If \code{FALSE}, error if a cached file with
the same \code{cache_file} name already exists.}
}
\description{
Cache a \code{ms_object} class object to a file
}
\keyword{internal}
