% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_column_definition.R
\name{copy_column_definition_list}
\alias{copy_column_definition_list}
\title{Create a new column definition based on an existing list}
\usage{
copy_column_definition_list(sp_list = NULL, ...)
}
\arguments{
\item{sp_list}{A \code{ms_list} object or a data frame created by
\code{\link[=get_sp_list_metadata]{get_sp_list_metadata()}}. Optional if additional parameters are provided to
\code{...} that can be used to get a SharePoint list to copy column definitions
from.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_list_metadata]{get_sp_list_metadata}}
  \describe{
    \item{\code{list_name,list_id}}{SharePoint List name or ID string.}
    \item{\code{keep}}{One of "all" (default), "editable", "external" (non-internal
fields). Argument determines if the returned list metadata includes read
only columns or hidden columns.}
    \item{\code{sync_fields}}{If \code{TRUE}, use the \code{sync_fields} method to sync the fields
of the local \code{ms_list} object with the fields of the SharePoint List source
before retrieving list metadata.}
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{drive}}{A \code{ms_drive} object. If \code{drive} is supplied, \code{drive_name} and
\code{drive_id} are ignored.}
    \item{\code{site_url}}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
    \item{\code{site}}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}
  }}
}
\description{
\code{\link[=copy_column_definition_list]{copy_column_definition_list()}} takes an existing SharePoint list and uses
the list metadata to create a column definition list that can be used to
create a new SharePoint list. NOTE: choice column types are not yet supported
so choice definitions from the source list are dropped.
}
\keyword{lists}
