% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_column_definition.R
\name{create_column_definition}
\alias{create_column_definition}
\alias{create_text_column}
\alias{create_choice_column}
\alias{create_number_column}
\alias{create_datetime_column}
\alias{create_boolean_column}
\alias{create_currency_column}
\alias{create_calculated_column}
\alias{create_lookup_column}
\alias{create_person_column}
\alias{create_group_column}
\alias{create_hyperlink_column}
\alias{create_picture_column}
\alias{create_thumbnail_column}
\alias{create_geolocation_column}
\alias{create_term_column}
\title{Create a column definition for use with the create column method for
SharePoint lists}
\usage{
create_column_definition(
  name,
  ...,
  .col_type = "text",
  enforce_unique = NULL,
  hidden = FALSE,
  deletable = NULL,
  indexed = NULL,
  sealed = NULL,
  propagate_changes = NULL,
  read_only = NULL,
  required = NULL,
  validation = NULL,
  default = get_column_default(),
  description = NULL,
  displayname = NULL,
  id = NULL
)

create_text_column(
  name,
  ...,
  multiple_lines = NULL,
  append_changes = NULL,
  lines = NULL,
  max_length = NULL,
  text_type = c("plain", "richText")
)

create_choice_column(
  name,
  choices,
  ...,
  allow_text = TRUE,
  display_as = c("dropDownMenu", "checkBoxes", "radioButtons"),
  allow_na = TRUE,
  na_replacement = "NA",
  split = NULL
)

create_number_column(
  name,
  ...,
  decimals = "automatic",
  display_as = NULL,
  max = NULL,
  min = NULL
)

create_datetime_column(
  name,
  ...,
  display_as = c("default", "friendly", "standard"),
  format = c("dateOnly", "dateTime")
)

create_boolean_column(name, ...)

create_currency_column(name, ..., locale = "en-us")

create_calculated_column(
  name,
  ...,
  formula,
  format = c("dateOnly", "dateTime"),
  output_type = c("text", "boolean", "currency", "dateTime", "number")
)

create_lookup_column(
  name,
  lookup_list_column,
  ...,
  lookup_list_id = NULL,
  lookup_list = NULL,
  allow_multiple = NULL,
  allow_unlimited_length = NULL,
  primary_lookup_column_id = NULL
)

create_person_column(
  name,
  ...,
  allow_multiple = NULL,
  display_as = NULL,
  from_type = "peopleOnly"
)

create_group_column(
  name,
  ...,
  allow_multiple = NULL,
  display_as = NULL,
  from_type = "peopleAndGroups"
)

create_hyperlink_column(name, ..., is_picture = FALSE)

create_picture_column(name, ..., is_picture = TRUE)

create_thumbnail_column(name, ...)

create_geolocation_column(name, ...)

create_term_column(name, ..., allow_multiple = TRUE, show_full_name = NULL)
}
\arguments{
\item{multiple_lines}{Logical. If \code{TRUE}, allow multiple lines of text.}

\item{append_changes}{Logical. If \code{TRUE}, append changes to existing value
for column.}

\item{lines}{Whole number.}

\item{max_length}{Whole number. Max length in number of characters.}

\item{text_type}{One of \code{c("plain", "richText")}}

\item{display_as}{Value displayed as option. For \code{create_choice_column} one
of\code{c("checkBoxes", "dropDownMenu", "radioButtons")}. For
\code{create_number_column}, one of \code{c("number", "percentage")}. For
\code{create_datetime_column}, one of \code{c("default", "friendly", "standard")}.}

\item{split}{
    character vector (or object which can be coerced to such)
    containing \link[base]{regular expression}(s) (unless \code{fixed = TRUE})
    to use for splitting.  If empty matches occur, in particular if
    \code{split} has length 0, \code{x} is split into single characters.
    If \code{split} has length greater than 1, it is re-cycled along
    \code{x}.
  }

\item{decimals}{One of \code{c("none", "one", "two", "three", "four", "five")} or
a numeric value between 0 and 5.}

\item{max, min}{Minimum and maximum values allowed in number column.}

\item{format}{\code{"dateOnly"} or \code{"dateTime"}. Required by
\code{create_calculated_column} if \code{output_type} is "dateTime" otherwise ignored.}

\item{locale}{Locale}

\item{formula}{Required string with formula for calculated column definition.
See \href{https://support.microsoft.com/en-us/office/examples-of-common-formulas-in-lists-d81f5f21-2b4e-45ce-b170-bf7ebf6988b3}{examples of common formulas in lists}.}

\item{output_type}{Value type returned by calculated formula. One of
\code{c("text", "boolean", "currency", "dateTime", "number")}}

\item{lookup_list_column}{Name of lookup column in the lookup list to use.}

\item{lookup_list_id, lookup_list}{Lookup list ID string or "ms_list" class
object with id value in list properties.}

\item{allow_multiple}{If \code{TRUE}, allow lookup column to return multiple
values. @param allow_unlimited_length If \code{TRUE}, allow lookup column to
return any length value. @param primary_lookup_column_id If column
definnition is for a secondary column, the primary lookup column ID must be
supplied.}

\item{from_type}{What type of resources to choose from. Defaults to
"peopleOnly" for \code{\link[=create_person_column]{create_person_column()}} or "peopleAndGroups" for
\code{\link[=create_group_column]{create_group_column()}}}

\item{is_picture}{Logical indicator for display of hyperlink value as link
(\code{FALSE}, default for \code{\link[=create_hyperlink_column]{create_hyperlink_column()}}) or image (\code{TRUE}, default
for \code{\link[=create_picture_column]{create_picture_column()}}).}
}
\description{
\code{\link[=create_column_definition]{create_column_definition()}} builds a named list with the properties of the
columnDefinition resource type.
}
\details{
\url{https://learn.microsoft.com/en-us/graph/api/resources/columndefinition?view=graph-rest-1.0}

Display as options

Display as options vary by columnDefinition type:
\itemize{
\item Options for personOrGroupColumn: https://learn.microsoft.com/en-us/graph/api/resources/personorgroupcolumn?view=graph-rest-1.0#displayas-options
}
}
\keyword{internal}
