% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_column_definition.R
\name{create_column_definition_list}
\alias{create_column_definition_list}
\title{Vectorized version of \code{create_column_definition()}}
\usage{
create_column_definition_list(name, .col_type = "text", ...)
}
\description{
Vectorized version of \code{create_column_definition()}
}
\keyword{internal}
