% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_column_definition.R
\name{create_column_definition_list}
\alias{create_column_definition_list}
\title{Vectorized version of \code{create_column_definition()}}
\usage{
create_column_definition_list(definitions, col_type = "text", ignore_na = TRUE)
}
\arguments{
\item{definitions}{A list or data frame with arguments to use in creation of column definitions.}
}
\description{
Vectorized version of \code{create_column_definition()}
}
\keyword{internal}
