% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{create_list_info}
\alias{create_list_info}
\title{Create listinfo object}
\usage{
create_list_info(
  template = c("documentLibrary", "genericList", "task", "survey", "announcements",
    "contacts"),
  content_types = NULL,
  hidden = NULL,
  call = caller_env()
)
}
\description{
Helper function for \code{\link[=create_sp_list]{create_sp_list()}}
}
\keyword{internal}
