% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{create_sp_list}
\alias{create_sp_list}
\alias{update_sp_list}
\alias{delete_sp_list}
\title{Create, update, or delete a SharePoint List}
\usage{
create_sp_list(
  list_name,
  ...,
  description = NULL,
  columns = NULL,
  template = "genericList",
  content_types = NULL,
  hidden = NULL,
  title_definition = list(required = FALSE),
  site_url = NULL,
  site = NULL,
  call = caller_env()
)

update_sp_list(
  list_name = NULL,
  list_id = NULL,
  sp_list = NULL,
  display_name = NULL,
  description = NULL,
  ...,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)

delete_sp_list(
  list_name = NULL,
  list_id = NULL,
  sp_list = NULL,
  confirm = TRUE,
  ...,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)
}
\arguments{
\item{list_name}{Required. List name used as \code{displayName} property.}

\item{...}{Additional parameters passed to \code{\link[=get_sp_site]{get_sp_site()}} or
\code{\link[Microsoft365R:get_sharepoint_site]{Microsoft365R::get_sharepoint_site()}}.}

\item{description}{Optional description.}

\item{columns}{Optional. Use \code{\link[=create_column_definition]{create_column_definition()}} to create a single
column definition or use \code{\link[=create_column_definition_list]{create_column_definition_list()}} to create a list
of column definitions.}

\item{template}{Type of template to use in creating the list.}

\item{content_types}{Optional. Set \code{TRUE} for \code{contentTypesEnabled} to be
enabled.}

\item{hidden}{Optional. Set \code{TRUE} for list to be hidden.}

\item{title_definition}{Named list used to update the column definition of
the default \code{"Title"} column created when using the \code{"genericList"} template.
By default, makes Title column optional.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{list_id}{List ID for list to update or delete.}

\item{sp_list}{A \code{Microsoft365R::ms_list} object.}

\item{display_name}{Display name to replace existing display name. Used by
\code{\link[=update_sp_list]{update_sp_list()}}.}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If \code{drive} is supplied, \code{drive_name} and
\code{drive_id} are ignored.}

\item{confirm}{If \code{TRUE}, confirm deletion of list before proceeding.}
}
\value{
For \code{\link[=create_sp_list]{create_sp_list()}}, invisibly returns a \code{ms_list} object for
the newly created list. For \code{\link[=update_sp_list]{update_sp_list()}}, the updated \code{ms_list}
object. For \code{\link[=delete_sp_list]{delete_sp_list()}}, invisibly returns \code{NULL}.
}
\description{
\code{\link[=create_sp_list]{create_sp_list()}} allows the creation of a SharePoint list for a site. See:
\url{https://learn.microsoft.com/en-us/graph/api/list-create?view=graph-rest-1.0&tabs=http}

\code{\link[=update_sp_list]{update_sp_list()}} allows the modification of the list display name and
description.

\code{\link[=delete_sp_list]{delete_sp_list()}} deletes an existing list and requires user confirmation by
default.

Notes on creating a SharePoint list:
\itemize{
\item Dashes (`"-"``) in list names are removed from the list name but retained
in the list display name.
\item If your definition includes calculated columns, these columns may need to
be added after the list is initially created using
\code{\link[=create_sp_list_column]{create_sp_list_column()}}.
}

Notes on updating a SharePoint list:
\itemize{
\item The \code{"Title"} column type is always a "text" type column and can't be
changed.
}
}
\keyword{lists}
