% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{create_sp_list}
\alias{create_sp_list}
\title{Create a SharePoint List}
\usage{
create_sp_list(
  list_name,
  ...,
  description = NULL,
  columns = NULL,
  template = "genericList",
  content_types = NULL,
  hidden = NULL,
  site_url = NULL,
  site = NULL,
  call = caller_env()
)
}
\arguments{
\item{list_name}{Required. List name used as \code{displayName} property.}

\item{...}{Additional parameters passed to \code{\link[=get_sp_site]{get_sp_site()}} or
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}.}

\item{description}{Optional description.}

\item{columns}{Optional. Use \code{\link[=create_column_definition]{create_column_definition()}} to create a single
column definition or use \code{\link[=create_column_definition_list]{create_column_definition_list()}} to create a list
of column definitions.}

\item{template}{Type of template to use in creating the list.}

\item{content_types}{Optional. Set \code{TRUE} for \code{contentTypesEnabled} to be enabled.}

\item{hidden}{Optional. Set \code{TRUE} for list to be hidden.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=create_sp_list]{create_sp_list()}} allows the creation of a SharePoint list for a site. See:
\url{https://learn.microsoft.com/en-us/graph/api/list-create?view=graph-rest-1.0&tabs=http}
}
\keyword{internal}
