% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{create_sp_list_column}
\alias{create_sp_list_column}
\alias{update_sp_list_column}
\alias{delete_sp_list_column}
\title{Create, update, and delete a SharePoint list column}
\usage{
create_sp_list_column(
  sp_list = NULL,
  ...,
  column_name = NULL,
  column_definition = NULL,
  list_name = NULL,
  site_url = NULL,
  site = NULL
)

update_sp_list_column(
  sp_list = NULL,
  column_name = NULL,
  column_id = NULL,
  ...,
  list_name = NULL,
  site_url = NULL,
  site = NULL,
  column_definition = NULL,
  column_name_type = "name"
)

delete_sp_list_column(
  sp_list = NULL,
  column_name = NULL,
  column_id = NULL,
  list_name = NULL,
  site_url = NULL,
  site = NULL,
  column_name_type = "name"
)
}
\arguments{
\item{sp_list}{A \code{ms_list} object. If supplied, \code{list_name}, \code{list_id},
\code{site_url}, \code{site}, \code{drive_name}, \code{drive_id}, \code{drive}, and any additional
parameters passed to \code{...} are all ignored.}

\item{...}{
  Arguments passed on to \code{\link[=create_column_definition]{create_column_definition}}
  \describe{
    \item{\code{name}}{Column name.}
    \item{\code{.col_type}}{Column type. Defaults to "text". Must be one of "boolean",
"calculated", "choice", "currency", "dateTime", "lookup", "number",
"personOrGroup", "text", "term", "hyperlinkOrPicture", "thumbnail",
"contentApprovalStatus", or "geolocation".}
    \item{\code{enforce_unique}}{Enforce unique values in column.}
    \item{\code{hidden}}{If \code{TRUE}, column will be hidden by default.}
    \item{\code{deletable}}{If \code{TRUE}, column can't be deleted separate from the list.}
    \item{\code{required}}{If \code{TRUE}, column will be required.}
    \item{\code{default}}{Default value set by helper \code{\link[=get_column_default]{get_column_default()}} function.}
    \item{\code{description}}{Column description.}
    \item{\code{displayname}}{Column display name.}
  }}

\item{column_name, column_id}{Column ID for column to delete.}

\item{column_definition}{List with column definition created with
\code{\link[=create_column_definition]{create_column_definition()}} or a related function. Optional if \code{column_name}
and any required additional parameters are provided.}

\item{list_name}{List name. Required if \code{sp_list} is \code{NULL}.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{column_name_type}{"name" or "displayName". Used to match column ID so
column_name must be unique if \code{column_name_type = "displayName"}.}
}
\description{
\code{\link[=create_sp_list_column]{create_sp_list_column()}} adds a column to a SharePoint list and
\code{\link[=delete_sp_list_column]{delete_sp_list_column()}} removes a column to a SharePoint list.
\code{\link[=update_sp_list_column]{update_sp_list_column()}} updates a column definition for an existing column
in a SharePoint list.
}
\details{
See documentation:
\url{https://learn.microsoft.com/en-us/graph/api/list-post-columns?view=graph-rest-1.0&tabs=http}
}
