% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{create_sp_list_column}
\alias{create_sp_list_column}
\alias{update_sp_list_column}
\alias{delete_sp_list_column}
\title{Create, update, and delete SharePoint list columns}
\usage{
create_sp_list_column(
  sp_list = NULL,
  ...,
  column_name = NULL,
  column_definition = NULL,
  list_name = NULL
)

update_sp_list_column(
  sp_list = NULL,
  column_name = NULL,
  column_id = NULL,
  ...,
  list_name = NULL,
  column_definition = NULL,
  column_name_type = "name"
)

delete_sp_list_column(
  sp_list = NULL,
  column_name = NULL,
  column_id = NULL,
  list_name = NULL,
  column_name_type = "name"
)
}
\arguments{
\item{sp_list}{A \code{ms_list} object. If supplied, \code{list_name}, \code{list_id},
\code{site_url}, \code{site}, \code{drive_name}, \code{drive_id}, \code{drive}, and any additional
parameters passed to \code{...} are all ignored.}

\item{...}{
  Arguments passed on to \code{\link[=create_column_definition]{create_column_definition}}
  \describe{
    \item{\code{}}{}
  }}

\item{column_definition}{List with column definition created with
\code{\link[=create_column_definition]{create_column_definition()}} or a related function. Optional if \code{column_name} and any required additional parameters are provided.}
}
\description{
\code{\link[=create_sp_list_column]{create_sp_list_column()}} adds a column to a SharePoint list and
\code{\link[=delete_sp_list_column]{delete_sp_list_column()}} removes a column to a SharePoint list.
\code{\link[=update_sp_list_column]{update_sp_list_column()}} updates a column definition for an existing column
in a SharePoint list (but is not yet implemented).
}
\details{
See documentation:
\url{https://learn.microsoft.com/en-us/graph/api/list-post-columns?view=graph-rest-1.0&tabs=http}
}
\keyword{internal}
