% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{create_sp_list_lookup_column}
\alias{create_sp_list_lookup_column}
\title{Create SharePoint list lookup column}
\usage{
create_sp_list_lookup_column(
  column_name,
  sp_list = NULL,
  sp_lookup_list,
  lookup_list_column = column_name,
  list_name = NULL,
  site = NULL,
  site_url = NULL,
  ...
)
}
\arguments{
\item{sp_list}{A \code{ms_list} object. If supplied, \code{list_name}, \code{site}, and
\code{site_url} are all ignored.}

\item{lookup_list_column}{Name of lookup column in the lookup list to use.}

\item{list_name}{List name. Required if \code{sp_list} is \code{NULL}.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{...}{
  Arguments passed on to \code{\link[=create_lookup_column]{create_lookup_column}}
  \describe{
    \item{\code{allow_multiple}}{If \code{TRUE}, allow lookup column to return multiple
values.}
    \item{\code{allow_unlimited_length}}{If \code{TRUE}, allow lookup column to
return any length value.}
    \item{\code{primary_lookup_column_id}}{If column definition is for a secondary
column, the primary lookup column ID must be supplied.}
  }}
}
\description{
\code{\link[=create_sp_list_lookup_column]{create_sp_list_lookup_column()}} is a wrapper for \code{\link[=create_lookup_column]{create_lookup_column()}}
and \code{\link[=create_sp_list_column]{create_sp_list_column()}} that allows for the creation of a column in one
list (the "lookup" list) and then create a lookup column in a second list.
}
\keyword{internal}
