% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_column_definition.R
\name{data_as_column_definition_list}
\alias{data_as_column_definition_list}
\title{Convert a data frame to a column definition list}
\usage{
data_as_column_definition_list(data, ..., split = "|", ignore_na = TRUE)
}
\arguments{
\item{data}{A data frame input. Column types are used to infer the
appropriate Microsoft Lists column definition.}

\item{ignore_na}{If \code{TRUE}, drop any parameters with a \code{NA} value.}
}
\description{
\code{\link[=data_as_column_definition_list]{data_as_column_definition_list()}} is used to create a column definition list
based on an existing data frame.
}
\examples{
data_as_column_definition_list(mtcars)

}
