% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item.R
\name{delete_sp_item}
\alias{delete_sp_item}
\title{Delete SharePoint items}
\usage{
delete_sp_item(
  path = NULL,
  ...,
  item_id = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  confirm = TRUE,
  by_item = FALSE,
  call = caller_env()
)
}
\arguments{
\item{path}{A SharePoint file URL or the relative path to a file located in a
SharePoint drive. If input is a relative path, the string should \emph{not}
include the drive name. If input is a shared file URL, the text "Shared "
is removed from the start of the SharePoint drive name by default. If file
is a document URL, the \code{.default_drive_name} argument is used as the
\code{drive_name} and the \code{item_id} is extracted from the URL.}

\item{...}{Additional parameters passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}} or \code{\link[=get_sp_site]{get_sp_site()}}.}

\item{item_id}{A SharePoint item ID passed to the \code{itemid} parameter of the
\code{get_item} method for \code{ms_drive} objects.}

\item{drive_name, drive_id}{SharePoint drive name or ID.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{confirm}{If \code{TRUE}, confirm before deleting item.}

\item{by_item}{For business OneDrive or SharePoint document libraries, you
may need to set \code{by_item = TRUE} to delete the contents of a folder
depending on the policies set up by your SharePoint administrator policies.
Note, that this method can be slow for large folders.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Delete SharePoint items
}
