% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item.R
\name{delete_sp_item}
\alias{delete_sp_item}
\title{Delete SharePoint items (files and directories)}
\usage{
delete_sp_item(
  path = NULL,
  confirm = TRUE,
  by_item = FALSE,
  ...,
  item_id = NULL,
  item = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  call = caller_env()
)
}
\arguments{
\item{path}{A SharePoint file URL or the relative path to a file located in a
SharePoint drive. If input is a relative path, the string should \emph{not}
include the drive name. If input is a shared file URL, the text "Shared "
is removed from the start of the SharePoint drive name by default. If file
is a document URL, the \code{default_drive_name} argument is used as the
\code{drive_name} and the \code{item_id} is extracted from the URL.}

\item{confirm}{If \code{TRUE}, confirm before deleting item. If \code{TRUE} and session
is not interactive, the function will error.}

\item{by_item}{For business OneDrive or SharePoint document libraries, you
may need to set \code{by_item = TRUE} to delete the contents of a folder
depending on the policies set up by your SharePoint administrator policies.
Note, that this method can be slow for large folders.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_item]{get_sp_item}}
  \describe{
    \item{\code{drive_name,drive_id}}{SharePoint drive name or ID.}
    \item{\code{as_data_frame}}{If \code{TRUE}, return a data frame. If \code{FALSE} (default),
return a \code{ms_item} or \code{ms_item_properties} object.}
  }}

\item{item_id}{A SharePoint item ID passed to the \code{itemid} parameter of the
\code{get_item} method for \code{ms_drive} objects.}

\item{item}{A \code{ms_drive_item} class object. Optional if path or other
parameters to get an SharePoint item are supplied.}

\item{drive_name, drive_id}{SharePoint drive name or ID.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=delete_sp_item]{delete_sp_item()}} deletes items including files and directories using the \code{delete} method for . By default
\code{confirm = TRUE}, which requires the user to respond to a prompt: "Do you
really want to delete the drive item ...? (yes/No/cancel)" to continue.
}
\details{
Trouble-shooting errors

If you get the error: "The resource you are attempting to access is locked",
you or another user may have the file or a file within the directory open for
editing. Close the file and try deleting the item again.

If you get the error: "Request was cancelled by event received. If attempting
to delete a non-empty folder, it's possible that it's on hold." set \code{by_item = TRUE} and try again.
}
