% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item.R
\name{download_sp_item}
\alias{download_sp_item}
\title{Download one or more items from SharePoint to a file}
\usage{
download_sp_item(
  file,
  path = "",
  drive_name = NULL,
  drive_id = NULL,
  site_url = NULL,
  dest = NULL,
  overwrite = FALSE,
  recursive = FALSE,
  parallel = FALSE,
  ...,
  call = caller_env()
)
}
\arguments{
\item{file}{Required. A SharePoint shared file URL, document URL, or, if
\code{item_id} is supplied, a file name to use with \code{path} to set \code{dest} with
location and filename for downloaded item.}

\item{path}{description}

\item{drive_name, drive_id}{Passed to \code{get_drive} method for SharePoint site
object.}

\item{site_url}{A SharePoint site URL in the format "https://\if{html}{\out{<tenant
name>}}.sharepoint.com/sites/\if{html}{\out{<site name>}}". A SharePoint file or document URL
will be parsed and a site URL built using the tenant and site name if
found.}

\item{dest, overwrite, recursive, parallel}{Parameters passed to \code{download}
method for \code{ms_item} object.}

\item{...}{Optional arguments that will ultimately be passed to \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token}}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=download_sp_item]{download_sp_item()}} wraps the \code{download} method for SharePoint items making
it easier to download items based on a shared file URL or document URL. The
default valye for \code{path} is \code{""} so, by default, SharePoint items are
downloaded to the current working directory. Set \code{overwrite = TRUE} to allow
this function to overwrite an existing file.
}
