% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item.R
\name{download_sp_item}
\alias{download_sp_item}
\alias{download_sp_file}
\title{Download one or more items from SharePoint to a file or folder}
\usage{
download_sp_item(
  path = NULL,
  new_path = "",
  ...,
  item_id = NULL,
  item = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  dest = NULL,
  overwrite = FALSE,
  recursive = FALSE,
  parallel = FALSE,
  call = caller_env()
)

download_sp_file(file, new_path = "", ..., call = caller_env())
}
\arguments{
\item{path, file}{Required. A SharePoint shared file URL, document URL, or, if
\code{item_id} is supplied, a file name to use with \code{path} to set \code{dest} with
location and filename for downloaded item.}

\item{new_path}{Path to directory for downloaded item. Optional if \code{dest} is
supplied.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_item]{get_sp_item}}
  \describe{
    \item{\code{drive_name,drive_id}}{SharePoint drive name or ID.}
    \item{\code{properties}}{If \code{TRUE}, use the \code{get_item_properties} method and return
item properties instead of the item.}
    \item{\code{as_data_frame}}{If \code{TRUE}, return a data frame. If \code{FALSE} (default),
return a \code{ms_item} or \code{ms_item_properties} object.}
  }}

\item{item_id}{A SharePoint item ID passed to the \code{itemid} parameter of the
\code{get_item} method for \code{ms_drive} objects.}

\item{item}{A \code{ms_drive_item} class object. Optional if path or other
parameters to get an SharePoint item are supplied.}

\item{drive_name, drive_id}{SharePoint drive name or ID.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{dest, overwrite, recursive, parallel}{Parameters passed to \code{download}
method for \code{ms_drive_item} object.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Invisibly returns the input \code{dest} or the \code{dest} parsed from the input
path or item properties.
}
\description{
\code{\link[=download_sp_item]{download_sp_item()}} wraps the \code{download} method for SharePoint items making
it easier to download items based on a shared file URL or document URL. The
default valye for \code{path} is \code{""} so, by default, SharePoint items are
downloaded to the current working directory. Set \code{overwrite = TRUE} to allow
this function to overwrite an existing file. \code{\link[=download_sp_file]{download_sp_file()}} is
identical except for the name of the path parameter (which is file instead of
path). Note, if the selected item is a folder and \code{recurse = TRUE}, it may
take some time to download the enclosed items and {Microsoft365R} does not
provide a progress bar for that operation.
}
