% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item.R
\name{download_sp_list}
\alias{download_sp_list}
\title{Download a SharePoint List}
\usage{
download_sp_list(
  ...,
  new_path = "",
  sp_list = NULL,
  fileext = "csv",
  keep_list_cols = c("createdBy", "lastModifiedBy"),
  call = caller_env()
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_sp_list]{get_sp_list}}
  \describe{
    \item{\code{list_name,list_id}}{SharePoint List name or ID string.}
    \item{\code{as_data_frame}}{If \code{TRUE}, return a data frame with a "ms_list" column.
\code{\link[=get_sp_list]{get_sp_list()}} returns a 1 row data frame and \code{\link[=list_sp_lists]{list_sp_lists()}} returns a
data frame with n rows or all lists available for the SharePoint site or
drive. Defaults to \code{FALSE}. Ignored is \code{metadata = TRUE} as list metadata
is always returned as a data frame.}
    \item{\code{metadata}}{If \code{TRUE}, \code{\link[=get_sp_list]{get_sp_list()}} applies the \code{get_column_info}
method to the returned SharePoint list and returns a data frame with column
metadata for the list.}
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{drive}}{A \code{ms_drive} object. If \code{drive} is supplied, \code{drive_name} and
\code{drive_id} are ignored.}
    \item{\code{site_url}}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}
    \item{\code{site}}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}
  }}

\item{new_path}{Optional path to new file. If not \code{new_path} provided, the
file name is pulled from the name of the SharePoint list using the provided
\code{fileext}. If \code{new_path} is provided, \code{fileext} is ignored.}

\item{sp_list}{SharePoint list object. If supplied, all parameters supplied
to \code{...} are ignored.}

\item{fileext}{File extension to use for output file. Must be \code{"csv"} or
\code{"xlsx"}.}

\item{keep_list_cols}{Column names for those columns to maintain in a list
format instead of attempting to convert to a character vector.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=download_sp_list]{download_sp_list()}} downloads a SharePoint list to a CSV or XLSX file.
}
