% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesAndDirectories.R
\name{downloadFileFromSharepoint}
\alias{downloadFileFromSharepoint}
\title{Download Sharepoint Item}
\usage{
downloadFileFromSharepoint(team_name, folder_name, file_name)
}
\arguments{
\item{team_name}{name of team or drive}

\item{folder_name}{replace back slashes with forward slashes `/`}

\item{file_name}{include extension, leave alone if you're looking for item id of a folder only}
}
\value{
full file path where the file is downloaded.
After use, consider deleting this file by `file.remove()` to keep temp directory clean.
}
\description{
Download Sharepoint Item
}
