% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_column_definition.R
\name{get_column_default}
\alias{get_column_default}
\title{Get a column default value or formula}
\usage{
get_column_default(
  value = NULL,
  formula = NULL,
  allow_null = TRUE,
  call = caller_env()
)
}
\arguments{
\item{value}{Value used as default value.}

\item{formula}{Formula used as default value.}

\item{allow_null}{If \code{TRUE}, return \code{NULL} if both \code{value} and \code{formula} are
\code{NULL}.}
}
\description{
\code{\link[=get_column_default]{get_column_default()}} returns a formula or value or NULL.
}
\keyword{lists}
