% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_group.R
\name{get_sp_group}
\alias{get_sp_group}
\alias{get_sp_site_group}
\alias{list_sp_group_members}
\title{Get SharePoint group for a site or list group members}
\usage{
get_sp_group(
  site_url = NULL,
  site_name = NULL,
  site_id = NULL,
  ...,
  site = NULL,
  call = caller_env()
)

list_sp_group_members(
  site_url = NULL,
  site_name = NULL,
  site_id = NULL,
  ...,
  sp_group = NULL,
  as_data_frame = TRUE,
  call = caller_env()
)
}
\arguments{
\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site_name, site_id}{Site name or ID of the SharePoint site as an
alternative to the SharePoint site URL. Exactly one of  \code{site_url},
\code{site_name}, and \code{site_id} must be supplied.}

\item{...}{Additional parameters passed to \code{\link[=get_sp_site]{get_sp_site()}} or
\code{\link[Microsoft365R:get_sharepoint_site]{Microsoft365R::get_sharepoint_site()}}.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{sp_group}{A "az_group" object returned by  \code{\link[=get_sp_group]{get_sp_group()}} or using the \code{get_group} method.}

\item{as_data_frame}{If \code{TRUE} (default), converted list of members into a
data frame with a list column named \code{az_user} that contains the member
list and properties converted into columns.}
}
\description{
\code{\link[=get_sp_group]{get_sp_group()}} gets the group associated with an individual site using the
\code{get_group} method. \code{\link[=list_sp_group_members]{list_sp_group_members()}} lists members in the group.
Note that, as of February 1, 2024, the returned member data frame when
\code{as_data_frame = TRUE} contains a large number of list columns that could be
coerced into character columns. This should be addressed in a later update.
}
\keyword{groups}
