% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item.R
\name{get_sp_item}
\alias{get_sp_item}
\alias{get_sp_drive}
\alias{get_sp_site}
\title{Get a SharePoint item based on a file URL or file name and site}
\usage{
get_sp_item(
  file = NULL,
  item_id = NULL,
  drive = NULL,
  drive_name = NULL,
  site_url = NULL,
  ...,
  .default = "Documents",
  call = caller_env()
)

get_sp_drive(
  drive_name = NULL,
  drive_id = NULL,
  site_url = NULL,
  site = NULL,
  ...,
  call = caller_env()
)

get_sp_site(site_url = NULL, ..., call = caller_env())
}
\arguments{
\item{file}{A SharePoint file URL or the relative path to a file located in a
SharePoint drive. If file is a relative path, the string should \emph{not}
include the drive name. If file is a shared file URL, the text "Shared " is
removed from the start of the SharePoint drive name by default. If file is
a document URL, the \code{.default} argument is used as the drive_name and
the \code{item_id} is extracted from the URL.}

\item{item_id}{A SharePoint item ID passed to the \code{itemid} parameter of the
\code{get_item} method for \code{ms_drive} objects.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{drive_name, drive_id}{Passed to \code{get_drive} method for SharePoint site
object.}

\item{site_url}{A SharePoint site URL in the format "https://\if{html}{\out{<tenant
name>}}.sharepoint.com/sites/\if{html}{\out{<site name>}}". A SharePoint file or document URL
will be parsed and a site URL built using the tenant and site name if
found.}

\item{...}{Optional arguments that will ultimately be passed to \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token}}.}

\item{.default}{Default drive name string used only if file is a document URL.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{site}{A \code{ms_site} object. If site is supplied, \code{site_url} and any
additional parameters passed to \code{...} are ignored.}
}
\description{
\code{\link[=get_sp_item]{get_sp_item()}} is a helper for the \code{{Microsoft365R}} methods that makes it
easier to get an \code{ms_item} object based on a file URL.
}
\keyword{internal}
