% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item.R
\name{get_sp_item}
\alias{get_sp_item}
\alias{get_sp_item_properties}
\title{Get a SharePoint item or item properties}
\usage{
get_sp_item(
  path = NULL,
  item_id = NULL,
  ...,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  properties = FALSE,
  as_data_frame = FALSE,
  call = caller_env()
)

get_sp_item_properties(
  path = NULL,
  item_id = NULL,
  ...,
  drive = NULL,
  drive_name = NULL,
  drive_id = NULL,
  site_url = NULL,
  as_data_frame = FALSE,
  call = caller_env()
)
}
\arguments{
\item{path}{A SharePoint file URL or the relative path to a file located in a
SharePoint drive. If input is a relative path, the string should \emph{not}
include the drive name. If input is a shared file URL, the text "Shared "
is removed from the start of the SharePoint drive name by default. If file
is a document URL, the \code{default_drive_name} argument is used as the
\code{drive_name} and the \code{item_id} is extracted from the URL.}

\item{item_id}{A SharePoint item ID passed to the \code{itemid} parameter of the
\code{get_item} method for \code{ms_drive} objects.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_drive]{get_sp_drive}}
  \describe{
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{default_drive_name}}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}
    \item{\code{cache}}{If \code{TRUE}, cache drive to a file using \code{\link[=cache_sp_drive]{cache_sp_drive()}}.}
    \item{\code{refresh}}{If \code{TRUE}, get a new drive even if the existing drive is
cached as a local option. If \code{FALSE}, use the cached \code{ms_drive} object if
it exists.}
    \item{\code{cache_file}}{File name for cached drive if \code{cache = TRUE}. Defaults to a
option set with \code{sharepointr.cache_file_drive} (which defaults to
\code{"sp_drive.rds"}).}
    \item{\code{site}}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}
    \item{\code{overwrite}}{If \code{TRUE}, replace the existing cached object named by
\code{cache_file} with the new object. If \code{FALSE}, error if a cached file with
the same \code{cache_file} name already exists.}
  }}

\item{drive_name, drive_id}{SharePoint drive name or ID.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{properties}{If \code{TRUE}, use the \code{get_item_properties} method and return
item properties instead of the item.}

\item{as_data_frame}{If \code{TRUE}, return a data frame. If \code{FALSE} (default),
return a \code{ms_item} or \code{ms_item_properties} object.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=get_sp_item]{get_sp_item()}} wraps the \code{get_item} method for \code{ms_drive} objects and
returns a \code{ms_drive_item} object by default. \code{\link[=get_sp_item_properties]{get_sp_item_properties()}} uses
the \code{get_item_properties} method (also available by setting \code{properties = TRUE} for \code{\link[=get_sp_item]{get_sp_item()}}).

Additional parameters in \code{...} are passed to \code{\link[=get_sp_drive]{get_sp_drive()}} by
\code{\link[=get_sp_item]{get_sp_item()}} or to \code{\link[=get_sp_item]{get_sp_item()}} by \code{\link[=get_sp_item_properties]{get_sp_item_properties()}} or
\code{\link[=delete_sp_item]{delete_sp_item()}}.
}
\seealso{
\link[Microsoft365R:ms_drive_item]{Microsoft365R::ms_drive_item}
}
