% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{get_sp_list_column}
\alias{get_sp_list_column}
\title{Get SharePoint list column definition for a single column}
\usage{
get_sp_list_column(
  sp_list = NULL,
  column_name = NULL,
  column_id = NULL,
  ...,
  list_name = NULL,
  site_url = NULL,
  site = NULL,
  column_name_type = "name",
  call = caller_env()
)
}
\arguments{
\item{sp_list}{A \code{ms_list} object. If supplied, \code{list_name}, \code{list_id},
\code{site_url}, \code{site}, \code{drive_name}, \code{drive_id}, \code{drive}, and any additional
parameters passed to \code{...} are all ignored.}

\item{column_name, column_id}{Column name or ID to get a definition for.}

\item{...}{
  Arguments passed on to \code{\link{get_sp_list}}
  \describe{
    \item{\code{list_name,list_id}}{SharePoint List name or ID string.}
    \item{\code{keep}}{One of "all" (default), "editable", "external" (non-internal
fields). Argument determines if the returned list metadata includes read
only columns or hidden columns.}
    \item{\code{sync_fields}}{If \code{TRUE}, use the \code{sync_fields} method to sync the fields
of the local \code{ms_list} object with the fields of the SharePoint List source
before retrieving list metadata.}
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{drive}}{A \code{ms_drive} object. If \code{drive} is supplied, \code{drive_name} and
\code{drive_id} are ignored.}
    \item{\code{filter}}{A string with \href{https://learn.microsoft.com/en-us/graph/query-parameters?tabs=http#filter-parameter}{an OData expression}
apply as a filter to the results. Learn more in the \href{https://learn.microsoft.com/en-us/graph/filter-query-parameter}{Microsoft Graph API documentation}
on using filter query parameters.}
    \item{\code{n}}{Maximum number of lists, plans, tasks, or other items to return.
Defaults to \code{NULL} which sets n to \code{Inf}.}
  }}

\item{list_name}{SharePoint List name. Optional if \code{sp_list} or \code{list_id} are provided.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{column_name_type}{"name" or "displayName". Used to match column ID so
column_name must be unique if \code{column_name_type = "displayName"}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=get_sp_list_column]{get_sp_list_column()}} gets a list column definition for a single column
specified by column name or ID.
}
\details{
See the \href{https://learn.microsoft.com/en-us/graph/api/columndefinition-get?view=graph-rest-1.0&tabs=http}{Get columnDefinition Graph API documentation} for more information.
}
