% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{get_sp_list_column}
\alias{get_sp_list_column}
\title{Get SharePoint list column definition}
\usage{
get_sp_list_column(
  sp_list = NULL,
  column_name = NULL,
  column_id = NULL,
  list_name = NULL,
  column_name_type = "name"
)
}
\arguments{
\item{sp_list}{A \code{ms_list} object. If supplied, \code{list_name}, \code{list_id},
\code{site_url}, \code{site}, \code{drive_name}, \code{drive_id}, \code{drive}, and any additional
parameters passed to \code{...} are all ignored.}
}
\description{
\code{\link[=get_sp_list_column]{get_sp_list_column()}} get a list column definition.
}
\details{
See Graph API documentation \url{https://learn.microsoft.com/en-us/graph/api/columndefinition-get?view=graph-rest-1.0&tabs=http}
}
