% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesAndDirectories.R
\name{getItemId}
\alias{getItemId}
\title{Sharepoint Item Id}
\usage{
getItemId(team_name, folder_path, filename = "")
}
\arguments{
\item{team_name}{name of team or drive}

\item{folder_path}{replace back slashes with forward slashes `/`. Feel free to keep spaces if present.}

\item{filename}{include extension, leave alone if you're looking for item id of a folder only}
}
\value{
item identification to be used for reading / writing / copying / moving / deleting
}
\description{
Sharepoint Item Id
}
