% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_drive.R
\name{list_sp_drives}
\alias{list_sp_drives}
\title{List drives for a SharePoint site}
\usage{
list_sp_drives(
  ...,
  site = NULL,
  filter = NULL,
  n = NULL,
  as_data_frame = TRUE,
  call = caller_env()
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_sp_site]{get_sp_site}}
  \describe{
    \item{\code{site_url}}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}
    \item{\code{site_name,site_id}}{Site name or ID of the SharePoint site as an
alternative to the SharePoint site URL. Exactly one of  \code{site_url},
\code{site_name}, and \code{site_id} must be supplied.}
    \item{\code{refresh}}{If \code{TRUE}, get a new site even if the existing site is cached
as a local option. If \code{FALSE}, use the cached \code{ms_site} object.}
    \item{\code{cache}}{If \code{TRUE}, cache site to a file using \code{\link[=cache_sp_site]{cache_sp_site()}}.}
    \item{\code{cache_file}}{File name for cached file if \code{cache = TRUE}. Defaults to
\code{"sp_site.rds"} or option set with \code{sharepointr.cache_file_site}.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
    \item{\code{overwrite}}{If \code{TRUE}, replace the existing cached object named by
\code{cache_file} with the new object. If \code{FALSE}, error if a cached file with
the same \code{cache_file} name already exists.}
  }}

\item{filter}{Filter to apply to query}

\item{n}{Max number of drives to return}
}
\description{
\code{\link[=list_sp_drives]{list_sp_drives()}} loads a SharePoint site uses the \code{list_drives} method to
returns a data frame with a list column or \code{ms_drive} objects or, if
\code{as_data_frame = FALSE}. This is helpful if a drive has been renamed and
can't easily be identified using a Drive URL alone.
}
\seealso{
\link[Microsoft365R:ms_site]{Microsoft365R::ms_site}
}
