% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_item_versions.R
\name{list_sp_item_versions}
\alias{list_sp_item_versions}
\title{List versions for a SharePoint item}
\usage{
list_sp_item_versions(..., sp_item = NULL, as_data_frame = TRUE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_sp_item]{get_sp_item}}
  \describe{
    \item{\code{path}}{A SharePoint file URL or the relative path to a file located in a
SharePoint drive. If input is a relative path, the string should \emph{not}
include the drive name. If input is a shared file URL, the text "Shared "
is removed from the start of the SharePoint drive name by default. If file
is a document URL, the \code{default_drive_name} argument is used as the
\code{drive_name} and the \code{item_id} is extracted from the URL.}
    \item{\code{item_id}}{A SharePoint item ID passed to the \code{itemid} parameter of the
\code{get_item} method for \code{ms_drive} objects.}
    \item{\code{item_url}}{A SharePoint item URL used to parse the item ID, drive name,
and site URL.}
    \item{\code{drive_name,drive_id}}{SharePoint drive name or ID.}
    \item{\code{drive}}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}
    \item{\code{properties}}{If \code{TRUE}, use the \code{get_item_properties} method and return
item properties instead of the item.}
    \item{\code{site_url}}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
  }}

\item{sp_item}{SharePoint item to get versions for. Additional parameters
passed to \code{...} are ignored if sp_item is supplied.}

\item{as_data_frame}{If \code{TRUE}, return a data frame of versions. If \code{FALSE},
return a list.}
}
\value{
A data frame if \code{as_data_frame = TRUE} or a list if \code{FALSE}.
}
\description{
This is an implementation of a new method for a \code{ms_item} object using the
\code{do_operation} method directly. The intent is to add a list_versions method
back to the Microsoft365R package which may include changes to the
functionality and output of this function.
}
\keyword{items}
