% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_user_info.R
\name{list_sp_site_user_info}
\alias{list_sp_site_user_info}
\title{List items from the hidden SharePoint "User Information List"}
\usage{
list_sp_site_user_info(..., sp_site = NULL, as_data_frame = TRUE)
}
\arguments{
\item{sp_site}{A \code{ms_site} object.}
}
\description{
List items from the hidden SharePoint "User Information List"
}
\keyword{internal}
