% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ms_obj_as_data_frame}
\alias{ms_obj_as_data_frame}
\title{Convert a ms_obj object to a data frame of properties with a list column of
objects}
\usage{
ms_obj_as_data_frame(
  ms_obj,
  obj_col = "ms_plan",
  recursive = FALSE,
  keep_list_cols = NULL,
  unlist_cols = TRUE,
  .name_repair = "universal_quiet",
  .error_call = caller_env()
)
}
\arguments{
\item{obj_col}{Column name for list column with \code{ms_} objects. Defaults to
\code{"ms_plan"}.}

\item{keep_list_cols}{Column names for those columns to maintain in a list
format instead of attempting to convert to a character vector.}
}
\description{
Convert a ms_obj object to a data frame of properties with a list column of
objects
}
\keyword{internal}
