% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sharepoint.R
\name{read_sharepoint}
\alias{read_sharepoint}
\title{Read a SharePoint item based on a file URL or file name and site
details}
\usage{
read_sharepoint(
  file,
  path = tempdir(),
  drive_name = NULL,
  drive_id = NULL,
  site_url = NULL,
  dest = NULL,
  overwrite = TRUE,
  recursive = FALSE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{file}{Required. A SharePoint shared file URL, document URL, or, if
\code{item_id} is supplied, a file name to use with \code{path} to set \code{dest} with
location and filename for downloaded item.}

\item{path}{description}

\item{drive_name, drive_id}{Passed to \code{get_drive} method for SharePoint site
object.}

\item{site_url}{A SharePoint site URL in the format "https://\if{html}{\out{<tenant
name>}}.sharepoint.com/sites/\if{html}{\out{<site name>}}". A SharePoint file or document URL
will be parsed and a site URL built using the tenant and site name if
found.}

\item{dest, overwrite, recursive, parallel}{Parameters passed to \code{download}
method for \code{ms_item} object.}

\item{...}{Optional arguments that will ultimately be passed to \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token}}.}
}
\description{
\code{\link[=read_sharepoint]{read_sharepoint()}} is designed to download a SharePoint item to a temporary
folder and read the file based on the file extension. The class of the
returned object depends on the file extension of the file parameter.
\itemize{
\item If the file is a csv or Microsoft Excel file (xlsx or xls), the
\code{\link[readr:read_delim]{readr::read_csv()}} or \code{\link[readxl:read_excel]{readxl::read_excel()}} is used to return a data frame.
\item If the file is a pptx or docx file, \code{\link[officer:read_docx]{officer::read_docx()}} or
\code{\link[officer:read_pptx]{officer::read_pptx()}} are used to read the file into a \code{rdocx} or \code{rpptx}
object.
\item If the file is a GeoPackage or GeoJSON file, \code{\link[sf:st_read]{sf::read_sf()}} is used to
read the file into a \code{sf} data frame.
}
}
