% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepointr-package.R
\docType{package}
\name{sharepointr-package}
\alias{sharepointr}
\alias{sharepointr-package}
\title{sharepointr: Read and Write from SharePoint Sites}
\description{
A R package extending Microsoft365R to make it easier to read, write, and work with SharePoint items, lists, and plans.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/sharepointr}
  \item \url{https://elipousson.github.io/sharepointr/}
  \item Report bugs at \url{https://github.com/elipousson/sharepointr/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]
}

}
\keyword{internal}
