% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_connection.R
\name{sp_connection}
\alias{sp_connection}
\title{Create a connection object to a SharePoint site}
\usage{
sp_connection(site, username, password, get_config = F, config)
}
\arguments{
\item{site}{character, URL of the SharePoint site e.g. https://org.sharepoint.com/sites/our_team}

\item{username}{character, an email address, ID or other}

\item{password}{character, a password for the associated user name}

\item{get_config}{boolean, whether to use a proxy script (from Internet Explorer settings).
Will fail if you are outside the network and not on the VPN}

\item{config}{httr::config request, a httr::config list}
}
\value{
a list with objects:
\itemize{
\item site - the input URL, a list with base and site parts
\item config - a list of proxy settings
\item cookie - a list of cookies
}
}
\description{
Create a connection object to a SharePoint site
}
\examples{
\dontrun{
 sp_con <- sp_connection(site = 'https://org.sharepoint.com/sites/our_team',
                         username = 'me@org.place.come',
                         password = 'mypass')
}
}
