% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dir.R
\name{sp_dir_create}
\alias{sp_dir_create}
\title{Create SharePoint folders}
\usage{
sp_dir_create(path, ..., drive_name = NULL, drive = NULL, call = caller_env())
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
\code{\link[=sp_dir_create]{sp_dir_create()}} is a wrapper for the \code{create_folder} method that handles
character vectors.
}
