% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dir.R
\name{sp_dir_create}
\alias{sp_dir_create}
\title{Create SharePoint folders}
\usage{
sp_dir_create(
  path,
  ...,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  relative = FALSE,
  call = caller_env()
)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_drive]{get_sp_drive}}
  \describe{
    \item{\code{default_drive_name}}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}
    \item{\code{cache}}{If \code{TRUE}, cache drive to a file using \code{\link[=cache_sp_drive]{cache_sp_drive()}}.}
    \item{\code{refresh}}{If \code{TRUE}, get a new drive even if the existing drive is
cached as a local option. If \code{FALSE}, use the cached \code{ms_drive} object if
it exists.}
    \item{\code{cache_file}}{File name for cached drive if \code{cache = TRUE}. Defaults to a
option set with \code{sharepointr.cache_file_drive} (which defaults to
\code{"sp_drive.rds"}).}
    \item{\code{site_url}}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}
    \item{\code{site}}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}
    \item{\code{overwrite}}{If \code{TRUE}, replace the existing cached object named by
\code{cache_file} with the new object. If \code{FALSE}, error if a cached file with
the same \code{cache_file} name already exists.}
  }}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If \code{drive} is supplied, \code{drive_name} and
\code{drive_id} are ignored.}

\item{relative}{If \code{TRUE} and \code{drive_name} is a folder URL, the values for
\code{path} are appended to the file path parsed from the url. If \code{relative} is
a character vector, it must be length 1 or the same length as path and
appended to path as a vector of parent directories. The second option takes
precedence over any file path parsed from the url.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=sp_dir_create]{sp_dir_create()}} is a wrapper for the \code{create_folder} method that handles
character vectors. If \code{drive_name} is a folder URL and \code{relative} is \code{TRUE},
the values for \code{path} are appended to the file path parsed from the url.
}
\examples{
drive_url <- "<link to SharePoint drive>"

if (is_sp_url(drive_url)) {
  sp_dir_create(
    path = "parent_folder/subfolder",
    drive_name = drive_url
  )

  sp_dir_create(
    path = c("subfolder1", "subfolder2", "subfolder3"),
    relative = "parent_folder",
    drive_name = drive_url
  )
}

dir_url <- "<link to SharePoint directory>"

if (is_sp_url(dir_url)) {
  sp_dir_create(
    path = c("subfolder1", "subfolder2", "subfolder3"),
    drive_name = dir_url,
    relative = TRUE
  )
}

}
