% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dir.R
\name{sp_dir_info}
\alias{sp_dir_info}
\title{List SharePoint files}
\usage{
sp_dir_info(
  path = NULL,
  ...,
  drive_name = NULL,
  drive = NULL,
  call = caller_env()
)
}
\arguments{
\item{path}{Path to directory or folder. SharePoint folder URLs are allowed.
If \code{NULL}, path is set to default "/".}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_drive]{get_sp_drive}}
  \describe{
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{site}}{A \code{ms_site} object. If site is supplied, \code{site_url} is ignored.}
    \item{\code{site_url}}{A SharePoint site URL in the format "https://\if{html}{\out{<tenant
name>}}.sharepoint.com/sites/\if{html}{\out{<site name>}}". Any SharePoint item or document
URL will be parsed and a site URL built using the tenant and site name if
found.}
  }}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=sp_dir_info]{sp_dir_info()}} is a wrapper for the \code{list_files} method.
}
