% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dir.R
\name{sp_dir_info}
\alias{sp_dir_info}
\alias{sp_dir_ls}
\title{List SharePoint files and folders}
\usage{
sp_dir_info(
  path = NULL,
  ...,
  info = "partial",
  full_names = TRUE,
  pagesize = 1000,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)

sp_dir_ls(
  path = NULL,
  ...,
  full_names = FALSE,
  pagesize = 1000,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)
}
\arguments{
\item{path}{Path to directory or folder. SharePoint folder URLs are allowed.
If \code{NULL}, path is set to default "/".}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_drive]{get_sp_drive}}
  \describe{
    \item{\code{.default_drive_name}}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}
    \item{\code{site_url}}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}
    \item{\code{site}}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}
  }}

\item{info}{The information to return: "partial", "name" or "all". If
"partial", a data frame is returned containing the name, size, ID and
whether the item is a file or folder. If "all", a data frame is returned containing all the
properties for each item (this can be large).}

\item{pagesize}{Maximum number of items to return. Defaults to 1000. Decrease
if you are experiencing timeouts.}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{full_name}{If \code{TRUE} (default), return the full file path as the name
for each item.}
}
\description{
\code{\link[=sp_dir_info]{sp_dir_info()}} is a wrapper for the \code{list_files} method.
}
