% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_drive.R
\name{sp_drive}
\alias{sp_drive}
\alias{get_sp_drive}
\alias{set_sp_drive}
\title{Get a SharePoint drive or set a default SharePoint drive}
\usage{
get_sp_drive(
  drive_name = NULL,
  drive_id = NULL,
  ...,
  site_url = NULL,
  site = NULL,
  .default_drive_name = getOption("sharepointr.default_drive_name", "Documents"),
  properties = FALSE,
  call = caller_env()
)

set_sp_drive(..., overwrite = FALSE, call = caller_env())
}
\arguments{
\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{...}{For \code{\link[=get_sp_drive]{get_sp_drive()}}, additional parameters passed to
\code{\link[=get_sp_site]{get_sp_site()}}. For \code{\link[=set_sp_drive]{set_sp_drive()}}, parameters passed to
\code{\link[=get_sp_drive]{get_sp_drive()}}.}

\item{site_url}{A SharePoint site URL in the format "https://\if{html}{\out{<tenant
name>}}.sharepoint.com/sites/\if{html}{\out{<site name>}}". Any SharePoint item or document
URL will be parsed and a site URL built using the tenant and site name if
found.}

\item{site}{A \code{ms_site} object. If site is supplied, \code{site_url} is ignored.}

\item{.default_drive_name}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}

\item{properties}{If \code{TRUE}, return the drive properties instead of the
\code{ms_drive} object.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{overwrite}{If \code{TRUE}, replace the existing option for
\code{sharepointr.default_drive}. If \code{FALSE} and the option is a \code{ms_drive}
object, return an error.}
}
\description{
\code{\link[=get_sp_drive]{get_sp_drive()}} wraps the \code{get_drive} method returns a \code{ms_drive} object.

\code{\link[=set_sp_drive]{set_sp_drive()}} allows you to set a default SharePoint drive for use by
other functions.
}
