% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{sp_list}
\alias{sp_list}
\alias{get_sp_list}
\alias{list_sp_lists}
\title{Get a SharePoint list, list items, or a list of SharePoint lists}
\usage{
get_sp_list(
  list_name = NULL,
  list_id = NULL,
  ...,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  items = TRUE,
  as_data_frame = TRUE,
  call = caller_env()
)

list_sp_lists(
  site_url = NULL,
  filter = NULL,
  n = NULL,
  ...,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  items = TRUE,
  as_data_frame = TRUE,
  call = caller_env()
)
}
\arguments{
\item{list_name, list_id}{SharePoint List name or ID string.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_drive]{get_sp_drive}}
  \describe{
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{.default_drive_name}}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}
    \item{\code{properties}}{If \code{TRUE}, return the drive properties instead of the
\code{ms_drive} object.}
  }}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{items}{If \code{TRUE} (default), use the \code{list_items} method to return a
data frame with all items in the specified list. If \code{FALSE}, return a
\code{ms_list} class object or, if \code{as_data_frame} is \code{TRUE}, a data frame with a "ms_list" column.}

\item{as_data_frame}{If \code{TRUE} (default), return a data frame with a
"ms_list" column. \code{\link[=get_sp_list]{get_sp_list()}} returns a 1 row data frame and
\code{\link[=list_sp_lists]{list_sp_lists()}} returns a data frame with n rows or all lists available
for the SharePoint site or drive.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{filter}{Filter for lists.}

\item{n}{Number of lists to return.}
}
\description{
\code{\link[=get_sp_list]{get_sp_list()}} is a wrapper for the \code{get_list} and \code{list_items} methods.
This function is still under development and does not support the URL parsing
used by \code{\link[=get_sp_item]{get_sp_item()}}. \code{\link[=list_sp_lists]{list_sp_lists()}} returns all lists for a SharePoint
site or drive as a list or data frame.
}
\seealso{
\itemize{
\item \link[Microsoft365R:ms_list]{Microsoft365R::ms_list}
\item \link[Microsoft365R:ms_list_item]{Microsoft365R::ms_list_item}
}
}
