% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list.R
\name{sp_list}
\alias{sp_list}
\alias{get_sp_list}
\alias{list_sp_lists}
\title{Get a SharePoint list or a list of SharePoint lists}
\usage{
get_sp_list(
  list_name = NULL,
  list_id = NULL,
  ...,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  metadata = FALSE,
  as_data_frame = TRUE,
  call = caller_env()
)

list_sp_lists(
  site_url = NULL,
  filter = NULL,
  n = Inf,
  ...,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  items = TRUE,
  as_data_frame = TRUE,
  call = caller_env()
)
}
\arguments{
\item{list_name, list_id}{SharePoint List name or ID string.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_drive]{get_sp_drive}}
  \describe{
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
    \item{\code{.default_drive_name}}{Drive name string used only if input is a document
URL and drive name is not part of the URL. Defaults to
\code{getOption("sharepointr.default_drive_name", "Documents")}}
  }}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{metadata}{If \code{TRUE}, \code{\link[=get_sp_list]{get_sp_list()}} applies the \code{get_column_info}
method to the returned SharePoint list and returns a data frame with column
metadata for the list.}

\item{as_data_frame}{If \code{TRUE} (default), return a data frame with a
"ms_list" column. \code{\link[=get_sp_list]{get_sp_list()}} returns a 1 row data frame and
\code{\link[=list_sp_lists]{list_sp_lists()}} returns a data frame with n rows or all lists available
for the SharePoint site or drive.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{filter}{A string with \href{https://learn.microsoft.com/en-us/graph/query-parameters?tabs=http#filter-parameter}{an OData expression}
apply as a filter to the results. Learn more in the \href{https://learn.microsoft.com/en-us/graph/filter-query-parameter}{Microsoft Graph API documentation}
on using filter query parameters.}

\item{n}{Maximum number of lists, plans, tasks, or other items to return.
Defaults to \code{NULL} which sets n to \code{Inf}.}
}
\value{
A data frame \code{as_data_frame = TRUE} or a \code{ms_list} object (or list
of \code{ms_list} objects) if \code{FALSE}.
}
\description{
\code{\link[=get_sp_list]{get_sp_list()}} is a wrapper for the \code{get_list} and \code{list_items} methods.
This function is still under development and does not support the URL parsing
used by \code{\link[=get_sp_item]{get_sp_item()}}. \code{\link[=list_sp_lists]{list_sp_lists()}} returns all lists for a SharePoint
site or drive as a list or data frame. Note, when using \code{filter} with
\code{\link[=get_sp_list]{get_sp_list()}}, names used in the expression must be prefixed with "fields/"
to distinguish them from item metadata.
}
\seealso{
\itemize{
\item \link[Microsoft365R:ms_list]{Microsoft365R::ms_list}
\item \link[Microsoft365R:ms_list_item]{Microsoft365R::ms_list_item}
}
}
