% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list_get_pipeline.R
\name{sp_list}
\alias{sp_list}
\title{Link a SharePoint connection to a list}
\usage{
sp_list(con, listName = NULL, listID = NULL)
}
\arguments{
\item{con}{A SharePoint connection returned
by sp_connection()}

\item{listName}{Name of the SharePoint list}
}
\description{
This method links a SharePoint connection to a
SharePoint list
}
\examples{
sp_con = sp_connection("https://yourdomain.sharepoint.com", "YourUsername", "YourPassword", Office365 = T)
sp_list = sp_list(sp_con, "yourList")
}
