% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list_item.R
\name{sp_list_item}
\alias{sp_list_item}
\alias{list_sp_list_items}
\alias{update_sp_list_item}
\alias{import_sp_list_items}
\alias{get_sp_list_item}
\title{Get, list, update, and delete SharePoint list items}
\usage{
list_sp_list_items(
  list_name = NULL,
  list_id = NULL,
  sp_list = NULL,
  ...,
  filter = NULL,
  select = NULL,
  all_metadata = FALSE,
  as_data_frame = TRUE,
  pagesize = 5000,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)

import_sp_list_items(
  data,
  list_name = NULL,
  list_id = NULL,
  id = NULL,
  ...,
  filter = NULL,
  select = NULL,
  all_metadata = FALSE,
  as_data_frame = TRUE,
  pagesize = 5000,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)

get_sp_list_item(
  id,
  list_name = NULL,
  ...,
  sp_list = NULL,
  list_id = NULL,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)

update_sp_list_item(
  id,
  list_name = NULL,
  ...,
  sp_list = NULL,
  list_id = NULL,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)
}
\arguments{
\item{list_name, list_id}{SharePoint List name or ID string.}

\item{...}{Additional parameters passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}} or \code{\link[=get_sp_site]{get_sp_site()}}.}

\item{filter}{A string with \href{https://learn.microsoft.com/en-us/graph/query-parameters?tabs=http#filter-parameter}{an OData expression}
apply as a filter to the results. Learn more in the \href{https://learn.microsoft.com/en-us/graph/filter-query-parameter}{Microsoft Graph API documentation}
on using filter query parameters.}

\item{select}{A character vector of column names to include in the returned
data frame of list items. If \code{NULL}, the data frame includes all columns
from the list.}

\item{all_metadata}{If \code{TRUE}, the returned data frame will contain extended
metadata as separate columns, while the data fields will be in a nested
data frame named fields. This is always set to \code{FALSE} if \code{n = NULL} or
\code{as_data_frame = FALSE}.}

\item{as_data_frame}{If \code{TRUE} (default), return a data frame with a
"ms_list" column. \code{\link[=get_sp_list]{get_sp_list()}} returns a 1 row data frame and
\code{\link[=list_sp_lists]{list_sp_lists()}} returns a data frame with n rows or all lists available
for the SharePoint site or drive.}

\item{pagesize}{Number of list items to return. Reduce from default of 5000
is experiencing timeouts.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=list_sp_list_items]{list_sp_list_items()}} lists \code{sp_list} items. Additional functions should be
completed for the \code{get_item}, \code{create_item}, \code{update_item}, and \code{delete_item}
methods documented in \link[Microsoft365R:ms_list]{Microsoft365R::ms_list}.
}
