% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_list_item.R
\name{sp_list_item}
\alias{sp_list_item}
\alias{list_sp_list_items}
\alias{get_sp_list_item}
\title{Get, list, update, and delete SharePoint list items}
\usage{
list_sp_list_items(
  list_name = NULL,
  list_id = NULL,
  sp_list = NULL,
  ...,
  filter = NULL,
  select = NULL,
  all_metadata = FALSE,
  as_data_frame = TRUE,
  display_nm = c("drop", "label", "replace"),
  name_repair = "unique",
  pagesize = 5000,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)

get_sp_list_item(
  id,
  list_name = NULL,
  list_id = NULL,
  sp_list = NULL,
  ...,
  site_url = NULL,
  site = NULL,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  call = caller_env()
)
}
\arguments{
\item{list_name, list_id}{SharePoint List name or ID string.}

\item{sp_list}{A \code{ms_list} object. If supplied, \code{list_name}, \code{list_id},
\code{site_url}, \code{site}, \code{drive_name}, \code{drive_id}, \code{drive}, and any additional
parameters passed to \code{...} are all ignored.}

\item{...}{
  Arguments passed on to \code{\link[=get_sp_list]{get_sp_list}}
  \describe{
    \item{\code{list_name,list_id}}{SharePoint List name or ID string.}
    \item{\code{metadata}}{If \code{TRUE}, \code{\link[=get_sp_list]{get_sp_list()}} applies the \code{get_column_info}
method to the returned SharePoint list and returns a data frame with column
metadata for the list.}
    \item{\code{drive_name,drive_id}}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}
  }}

\item{filter}{A string with \href{https://learn.microsoft.com/en-us/graph/query-parameters?tabs=http#filter-parameter}{an OData expression}
apply as a filter to the results. Learn more in the \href{https://learn.microsoft.com/en-us/graph/filter-query-parameter}{Microsoft Graph API documentation}
on using filter query parameters.}

\item{select}{A character vector of column names to include in the returned
data frame of list items. If \code{NULL}, the data frame includes all columns
from the list.}

\item{all_metadata}{If \code{TRUE}, the returned data frame will contain extended
metadata as separate columns, while the data fields will be in a nested
data frame named fields. This is always set to \code{FALSE} if \code{n = NULL} or
\code{as_data_frame = FALSE}.}

\item{as_data_frame}{If \code{TRUE}, return a data frame with a "ms_list" column.
\code{\link[=get_sp_list]{get_sp_list()}} returns a 1 row data frame and \code{\link[=list_sp_lists]{list_sp_lists()}} returns a
data frame with n rows or all lists available for the SharePoint site or
drive. Defaults to \code{FALSE}. Ignored is \code{metadata = TRUE} as list metadata
is always returned as a data frame.}

\item{display_nm}{Option of "drop" (default), "label", or "replace". If
"drop", display names are not accessed or used. If "label", display names
are used to label matching columns in the returned data frame. If
"replace", display names replace column names in the returned data frame.
When working with the last option, the \code{name_repair} argument is required
since there is no requirement on SharePoint for lists to use unique display
names and invalid data frames can result.}

\item{name_repair}{Passed to repair argument of \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}}

\item{pagesize}{Number of list items to return. Reduce from default of 5000
is experiencing timeouts.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If \code{drive} is supplied, \code{drive_name} and
\code{drive_id} are ignored.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{id}{Required. A SharePoint list item ID typically an integer for the
record number starting from 1 with the first record.}
}
\description{
\code{\link[=list_sp_list_items]{list_sp_list_items()}} lists \code{sp_list} items. Additional functions should be
completed for the \code{get_item}, \code{create_item}, \code{update_item}, and \code{delete_item}
methods documented in \link[Microsoft365R:ms_list]{Microsoft365R::ms_list}.
}
