% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_plan.R
\name{sp_plan}
\alias{sp_plan}
\alias{get_sp_plan}
\alias{list_sp_plans}
\title{Get a SharePoint plan or list SharePoint plans}
\usage{
get_sp_plan(
  plan_title = NULL,
  plan_id = NULL,
  ...,
  site = NULL,
  site_url = NULL,
  call = caller_env()
)

list_sp_plans(
  ...,
  filter = NULL,
  n = NULL,
  site = NULL,
  as_data_frame = TRUE,
  call = caller_env()
)
}
\arguments{
\item{plan_title, plan_id}{Planner title or ID. Exactly one of the two
arguments must be supplied.}

\item{...}{Additional arguments passed to \code{\link[=get_sp_site_group]{get_sp_site_group()}}.}

\item{site}{A \code{ms_site} object. If \code{site} is supplied, \code{site_url},
\code{site_name}, and \code{site_id} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL can also be parsed to build a site URL using the tenant and site name
included in the URL.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{filter}{Filter for plans.}

\item{n}{Number of plans to return.}

\item{as_data_frame}{If \code{TRUE} (default), return a data frame with the plan
tite, id, creation date/time, and owner ID with a list column named
"ms_plan" containing \code{ms_plan} objects. If \code{FALSE}, return a list of
\code{ms_plan} objects.}
}
\value{
For \code{\link[=get_sp_plan]{get_sp_plan()}}, a \code{ms_plan} class object.

For \code{\link[=list_sp_plans]{list_sp_plans()}}, A list of \code{ms_plan} class objects or a data
frame with a list column named "ms_plan".
}
\description{
\code{\link[=get_sp_plan]{get_sp_plan()}} returns a single \code{ms_plan} object using the \code{get_plan}
method. \code{\link[=list_sp_plans]{list_sp_plans()}} returns a data frame with plan properties or a list
of \code{ms_plan} objects.
}
\seealso{
\link[Microsoft365R:ms_plan]{Microsoft365R::ms_plan}
}
