% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_site.R
\name{sp_site}
\alias{sp_site}
\alias{get_sp_site}
\alias{set_sp_site}
\title{Get a SharePoint site or set a default SharePoint site}
\usage{
get_sp_site(
  site_url = NULL,
  site_name = NULL,
  site_id = NULL,
  ...,
  call = caller_env()
)

set_sp_site(..., overwrite = FALSE, call = caller_env())
}
\arguments{
\item{site_url}{A SharePoint site URL in the format "https://\if{html}{\out{<tenant
name>}}.sharepoint.com/sites/\if{html}{\out{<site name>}}". Any SharePoint item or document
URL will be parsed and a site URL built using the tenant and site name if
found.}

\item{site_name, site_id}{Site name or ID of the SharePoint site as an
alternative to the SharePoint site URL. Only one of the three arguments
should be supplied.}

\item{...}{
  Arguments passed on to \code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site}}
  \describe{
    \item{\code{app}}{A custom app registration ID to use for authentication. See below.}
    \item{\code{scopes}}{The Microsoft Graph scopes (permissions) to obtain. It should never be necessary to change these.}
    \item{\code{token}}{An AAD OAuth token object, of class \code{AzureAuth::AzureToken}. If supplied, the \code{tenant}, \code{app}, \code{scopes} and \code{...} arguments will be ignored. See "Authenticating with a token" below.}
    \item{\code{tenant}}{For \code{get_business_onedrive}, \code{get_sharepoint_site} and \code{get_team}, the name of your Azure Active Directory (AAD) tenant. If not supplied, use the value of the \code{CLIMICROSOFT365_TENANT} environment variable, or "common" if that is unset.}
    \item{\code{site_name,site_url,site_id}}{For \code{get_sharepoint_site}, either the name, web URL or ID of the SharePoint site to retrieve. Supply exactly one of these.}
  }}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{overwrite}{If \code{TRUE}, replace the existing option for
\code{sharepointr.default_site}. If \code{FALSE} and the option is a \code{ms_site}
object, return an error.}
}
\description{
\code{\link[=get_sp_site]{get_sp_site()}} is a wrapper for \code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}} and
returns a \code{ms_site} object. \code{\link[=set_sp_site]{set_sp_site()}} allows you to set a default
SharePoint site for use by other functions.
}
