% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_url_parse.R
\name{sp_url_parse}
\alias{sp_url_parse}
\alias{sp_url_parse_hostname}
\alias{sp_url_parse_path}
\alias{sp_url_parse_query}
\title{Parse a SharePoint URL using \code{httr2::url_parse}}
\usage{
sp_url_parse(url, call = caller_env())

sp_url_parse_hostname(hostname, tenant = "[a-zA-Z0-9.-]+", scheme = "https")

sp_url_parse_path(
  path,
  url_type = "w|x|p|o|b|t|i|v|f|u|li",
  permissions = "r|s|t|u|g",
  drive_name_prefix = "Shared ",
  .default_drive_name = "Documents"
)

sp_url_parse_query(query)
}
\description{
Parse a SharePoint URL using \code{httr2::url_parse}
}
\keyword{internal}
