% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sharepoint.R
\name{write_sharepoint}
\alias{write_sharepoint}
\title{Write an object to file and upload file to SharePoint}
\usage{
write_sharepoint(
  x,
  file,
  dest,
  ...,
  new_path = tempdir(),
  overwrite = FALSE,
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  site_name = NULL,
  site_id = NULL,
  site = NULL,
  blocksize = 327680000,
  recursive = FALSE,
  parallel = FALSE,
  call = caller_env()
)
}
\arguments{
\item{x}{Object to write to file and upload to SharePoint.}

\item{file}{Path for file or directory to upload. Optional if \code{src} is
supplied.}

\item{dest}{Destination on SharePoint for file to upload. SharePoint folder
URLs are supported.}

\item{...}{Additional parameters passed to write function.}

\item{new_path}{Path to write file to. Defaults to \code{\link[=tempdir]{tempdir()}}}

\item{overwrite}{If \code{FALSE} (default), error if an item with the name
specified in \code{file} or \code{src} already exists at the specified destination.
If \code{TRUE}, overwrite any existing items with the same name. The latter is
the default for the \code{upload_file} method.}

\item{drive_name, drive_id}{SharePoint Drive name or ID passed to \code{get_drive}
method for SharePoint site object.}

\item{drive}{A \code{ms_drive} object. If drive is supplied, \code{drive_name},
\code{site_url}, and any additional parameters passed to \code{...} are ignored.}

\item{site_url}{A SharePoint site URL in the format "https://[tenant
name].sharepoint.com/sites/[site name]". Any SharePoint item or document
URL will be parsed and a site URL built using the tenant and site name if
found.}

\item{site_name, site_id}{Site name or ID of the SharePoint site as an
alternative to the SharePoint site URL. Only one of the three arguments
should be supplied.}

\item{site}{A \code{ms_site} object. If site is supplied, \code{site_url} is ignored.}

\item{blocksize, recursive, parallel}{Additional parameters passed to
\code{upload_folder} or \code{upload_file} method for \code{ms_drive} objects.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=write_sharepoint]{write_sharepoint()}} uses \code{\link[=upload_sp_item]{upload_sp_item()}} to upload an object created
using:
\itemize{
\item \code{\link[sf:st_write]{sf::write_sf()}} if x is an \code{sf} object
\item \code{\link[readr:write_delim]{readr::write_csv()}} if x is a \code{data.frame}
\item The \code{\link[=print]{print()}} method from \code{{officer}} if x is a \code{rdocx}, \code{rpptx}, or \code{rxlsx} object
\item \code{\link[readr:read_rds]{readr::write_rds()}} if x is any other class
}
}
